Refine local copy module decomposition
