+++ crates/engine/src/local_copy/executor/cleanup.rs
@@ -1,0 +1,166 @@
+//! Deletion helpers for extraneous or source entries.
+
+use std::collections::HashSet;
+use std::ffi::OsString;
+use std::fs;
+use std::io;
+use std::path::{Path, PathBuf};
+use std::time::Duration;
+
+use crate::local_copy::{CopyContext, LocalCopyAction, LocalCopyError, LocalCopyRecord};
+
+pub(crate) fn delete_extraneous_entries(
+    context: &mut CopyContext,
+    destination: &Path,
+    relative: Option<&Path>,
+    source_entries: &[OsString],
+) -> Result<(), LocalCopyError> {
+    let mut skipped_due_to_limit = 0u64;
+    let mut keep = HashSet::with_capacity(source_entries.len());
+    for name in source_entries {
+        keep.insert(name.clone());
+    }
+
+    let read_dir = match fs::read_dir(destination) {
+        Ok(iter) => iter,
+        Err(error) if error.kind() == io::ErrorKind::NotFound => return Ok(()),
+        Err(error) => {
+            return Err(LocalCopyError::io(
+                "read destination directory",
+                destination.to_path_buf(),
+                error,
+            ));
+        }
+    };
+
+    for entry in read_dir {
+        context.enforce_timeout()?;
+        let entry = entry.map_err(|error| {
+            LocalCopyError::io("read destination entry", destination.to_path_buf(), error)
+        })?;
+        let name = entry.file_name();
+
+        if keep.contains(&name) {
+            continue;
+        }
+
+        let name_path = PathBuf::from(name.as_os_str());
+        let path = destination.join(&name_path);
+        let entry_relative = match relative {
+            Some(base) => base.join(&name_path),
+            None => name_path.clone(),
+        };
+
+        let file_type = entry.file_type().map_err(|error| {
+            LocalCopyError::io("inspect extraneous destination entry", path.clone(), error)
+        })?;
+
+        if !context.allows_deletion(entry_relative.as_path(), file_type.is_dir()) {
+            continue;
+        }
+
+        if let Some(limit) = context.options().max_deletion_limit() {
+            if context.summary().items_deleted() >= limit {
+                skipped_due_to_limit = skipped_due_to_limit.saturating_add(1);
+                continue;
+            }
+        }
+
+        if context.mode().is_dry_run() {
+            context.summary_mut().record_deletion();
+            context.record(LocalCopyRecord::new(
+                entry_relative,
+                LocalCopyAction::EntryDeleted,
+                0,
+                None,
+                Duration::default(),
+                None,
+            ));
+            context.register_progress();
+            continue;
+        }
+
+        context.backup_existing_entry(&path, Some(entry_relative.as_path()), file_type)?;
+        remove_extraneous_path(path, file_type)?;
+        context.summary_mut().record_deletion();
+        context.record(LocalCopyRecord::new(
+            entry_relative,
+            LocalCopyAction::EntryDeleted,
+            0,
+            None,
+            Duration::default(),
+            None,
+        ));
+        context.register_progress();
+    }
+
+    if skipped_due_to_limit > 0 {
+        return Err(LocalCopyError::delete_limit_exceeded(skipped_due_to_limit));
+    }
+
+    Ok(())
+}
+
+fn remove_extraneous_path(path: PathBuf, file_type: fs::FileType) -> Result<(), LocalCopyError> {
+    let context = if file_type.is_dir() {
+        "remove extraneous directory"
+    } else {
+        "remove extraneous entry"
+    };
+
+    let result = if file_type.is_dir() {
+        fs::remove_dir_all(&path)
+    } else {
+        fs::remove_file(&path)
+    };
+
+    match result {
+        Ok(()) => Ok(()),
+        Err(error) if error.kind() == io::ErrorKind::NotFound => Ok(()),
+        Err(error) => Err(LocalCopyError::io(context, path, error)),
+    }
+}
+
+pub(crate) fn remove_source_entry_if_requested(
+    context: &mut CopyContext,
+    source: &Path,
+    record_path: Option<&Path>,
+    file_type: fs::FileType,
+) -> Result<(), LocalCopyError> {
+    if !context.remove_source_files_enabled() || context.mode().is_dry_run() {
+        return Ok(());
+    }
+
+    let source_type = match fs::symlink_metadata(source) {
+        Ok(metadata) => metadata.file_type(),
+        Err(_) => file_type,
+    };
+
+    if source_type.is_dir() {
+        return Ok(());
+    }
+
+    match fs::remove_file(source) {
+        Ok(()) => {
+            context.summary_mut().record_source_removed();
+            if let Some(path) = record_path {
+                context.record(LocalCopyRecord::new(
+                    path.to_path_buf(),
+                    LocalCopyAction::SourceRemoved,
+                    0,
+                    None,
+                    Duration::default(),
+                    None,
+                ));
+            }
+            context.register_progress();
+            Ok(())
+        }
+        Err(error) if error.kind() == io::ErrorKind::NotFound => Ok(()),
+        Err(error) => Err(LocalCopyError::io(
+            "remove source entry",
+            source.to_path_buf(),
+            error,
+        )),
+    }
+}
