+++ crates/engine/src/local_copy/executor/util.rs
@@ -1,0 +1,19 @@
+//! Shared filesystem helpers for local copy execution.
+
+use std::fs;
+use std::path::Path;
+
+use crate::local_copy::LocalCopyError;
+
+pub(crate) fn non_empty_path(path: &Path) -> Option<&Path> {
+    if path.as_os_str().is_empty() {
+        None
+    } else {
+        Some(path)
+    }
+}
+
+pub(crate) fn follow_symlink_metadata(path: &Path) -> Result<fs::Metadata, LocalCopyError> {
+    fs::metadata(path)
+        .map_err(|error| LocalCopyError::io("inspect symlink target", path.to_path_buf(), error))
+}
