+++ crates/engine/src/local_copy/mod.rs
@@ -52,7 +52,11 @@
 mod context;
 mod dir_merge;
 mod error;
+mod executor;
 mod filter_program;
+mod hard_links;
+mod metadata_sync;
+mod operands;
 mod options;
 mod overrides;
 mod plan;
@@ -83,60 +87,27 @@ pub(crate) use dir_merge::{
     resolve_dir_merge_path,
 };
 
-use std::cell::Cell;
-use std::cmp::Ordering;
-#[cfg(unix)]
-use std::collections::HashMap;
-use std::collections::HashSet;
-use std::ffi::{OsStr, OsString};
-use std::fs;
-use std::io::{self, Read, Seek, SeekFrom, Write};
-use std::num::NonZeroU8;
-use std::path::{Component, Path, PathBuf};
-use std::process;
-use std::sync::atomic::{AtomicUsize, Ordering as AtomicOrdering};
-use std::time::{Duration, Instant, SystemTime};
+pub(crate) use executor::*;
 
-#[cfg(unix)]
-use rustix::{
-    fd::AsFd,
-    fs::{FallocateFlags, fallocate},
-    io::Errno,
-};
+pub(crate) use hard_links::HardLinkTracker;
 
-use rsync_checksums::strong::{Md4, Md5, Xxh3, Xxh3_128, Xxh64};
+pub(crate) use metadata_sync::map_metadata_error;
 #[cfg(feature = "acl")]
-use rsync_meta::sync_acls;
+pub(crate) use metadata_sync::sync_acls_if_requested;
 #[cfg(feature = "xattr")]
-use rsync_meta::sync_xattrs;
-use rsync_meta::{
-    MetadataError, MetadataOptions, apply_directory_metadata_with_options,
-    apply_file_metadata_with_options, apply_symlink_metadata_with_options, create_device_node,
-    create_fifo,
-};
-use rsync_protocol::ProtocolVersion;
+pub(crate) use metadata_sync::sync_xattrs_if_requested;
 
-use crate::delta::{DeltaSignatureIndex, SignatureLayoutParams, calculate_signature_layout};
-use crate::signature::{SignatureAlgorithm, SignatureError, generate_file_signature};
+pub(crate) use operands::{DestinationSpec, SourceSpec, operand_is_remote};
+
+pub use filter_program::{
+    DirMergeEnforcedKind, DirMergeOptions, DirMergeRule, ExcludeIfPresentRule, FilterProgram,
+    FilterProgramEntry, FilterProgramError,
+};
+
+use std::sync::atomic::AtomicUsize;
+
 const COPY_BUFFER_SIZE: usize = 128 * 1024;
 static NEXT_TEMP_FILE_ID: AtomicUsize = AtomicUsize::new(0);
-
-use overrides::{create_hard_link, device_identifier};
-#[cfg(test)]
-pub(crate) fn with_hard_link_override<F, R>(override_fn: F, action: impl FnOnce() -> R) -> R
-where
-    F: Fn(&Path, &Path) -> io::Result<()> + 'static,
-{
-    overrides::with_hard_link_override(override_fn, action)
-}
-
-#[cfg(test)]
-pub(crate) fn with_device_id_override<F, R>(override_fn: F, action: impl FnOnce() -> R) -> R
-where
-    F: Fn(&Path, &fs::Metadata) -> Option<u64> + 'static,
-{
-    overrides::with_device_id_override(override_fn, action)
-}
 
 #[cfg(unix)]
 const CROSS_DEVICE_ERROR_CODE: i32 = 18;
@@ -147,3652 +118,20 @@ const CROSS_DEVICE_ERROR_CODE: i32 = 17;
 #[cfg(not(any(unix, windows)))]
 const CROSS_DEVICE_ERROR_CODE: i32 = 18;
 
-pub use filter_program::{
-    DirMergeEnforcedKind, DirMergeOptions, DirMergeRule, ExcludeIfPresentRule, FilterProgram,
-    FilterProgramEntry, FilterProgramError,
-};
-
-#[cfg(unix)]
-#[derive(Default)]
-struct HardLinkTracker {
-    entries: HashMap<HardLinkKey, PathBuf>,
-}
-
-#[cfg(unix)]
-#[derive(Clone, Copy, Debug, Eq, Hash, PartialEq)]
-struct HardLinkKey {
-    device: u64,
-    inode: u64,
-}
-
-#[cfg(unix)]
-impl HardLinkTracker {
-    fn new() -> Self {
-        Self {
-            entries: HashMap::new(),
-        }
-    }
-
-    fn existing_target(&self, metadata: &fs::Metadata) -> Option<PathBuf> {
-        Self::key(metadata).and_then(|key| self.entries.get(&key).cloned())
-    }
-
-    fn record(&mut self, metadata: &fs::Metadata, destination: &Path) {
-        if let Some(key) = Self::key(metadata) {
-            self.entries.insert(key, destination.to_path_buf());
-        }
-    }
-
-    fn key(metadata: &fs::Metadata) -> Option<HardLinkKey> {
-        use std::os::unix::fs::MetadataExt;
-
-        if metadata.nlink() > 1 {
-            Some(HardLinkKey {
-                device: metadata.dev(),
-                inode: metadata.ino(),
-            })
-        } else {
-            None
-        }
-    }
-}
-
-#[cfg(not(unix))]
-#[derive(Default)]
-struct HardLinkTracker;
-
-#[cfg(not(unix))]
-impl HardLinkTracker {
-    const fn new() -> Self {
-        Self
-    }
-
-    fn existing_target(&self, _metadata: &fs::Metadata) -> Option<PathBuf> {
-        None
-    }
-
-    fn record(&mut self, _metadata: &fs::Metadata, _destination: &Path) {}
-}
-
-#[cfg(feature = "xattr")]
-fn sync_xattrs_if_requested(
-    preserve_xattrs: bool,
-    mode: LocalCopyExecution,
-    source: &Path,
-    destination: &Path,
-    follow_symlinks: bool,
-) -> Result<(), LocalCopyError> {
-    if preserve_xattrs && !mode.is_dry_run() {
-        sync_xattrs(source, destination, follow_symlinks).map_err(map_metadata_error)?;
-    }
-    Ok(())
-}
-
-#[cfg(feature = "acl")]
-fn sync_acls_if_requested(
-    preserve_acls: bool,
-    mode: LocalCopyExecution,
-    source: &Path,
-    destination: &Path,
-    follow_symlinks: bool,
-) -> Result<(), LocalCopyError> {
-    if preserve_acls && !mode.is_dry_run() {
-        sync_acls(source, destination, follow_symlinks).map_err(map_metadata_error)?;
-    }
-    Ok(())
-}
-
-/// Source operand within a [`LocalCopyPlan`].
-#[derive(Clone, Debug, Eq, PartialEq)]
-pub struct SourceSpec {
-    path: PathBuf,
-    copy_contents: bool,
-    relative_prefix_components: Option<usize>,
-}
-
-impl SourceSpec {
-    fn from_operand(operand: &OsString) -> Result<Self, LocalCopyError> {
-        if operand.is_empty() {
-            return Err(LocalCopyError::invalid_argument(
-                LocalCopyArgumentError::EmptySourceOperand,
-            ));
-        }
-
-        if operand_is_remote(operand.as_os_str()) {
-            return Err(LocalCopyError::invalid_argument(
-                LocalCopyArgumentError::RemoteOperandUnsupported,
-            ));
-        }
-
-        let copy_contents = has_trailing_separator(operand.as_os_str());
-        Ok(Self {
-            path: PathBuf::from(operand),
-            copy_contents,
-            relative_prefix_components: detect_relative_prefix_components(operand.as_os_str()),
-        })
-    }
-
-    fn relative_root(&self) -> Option<PathBuf> {
-        use std::path::Component;
-
-        let skip = self.relative_prefix_components.unwrap_or(0);
-        let mut index = 0;
-        let mut relative = PathBuf::new();
-
-        for component in self.path.components() {
-            if index < skip {
-                index += 1;
-                continue;
-            }
-
-            index += 1;
-
-            match component {
-                Component::CurDir | Component::RootDir => {}
-                Component::Prefix(prefix) => {
-                    relative.push(Path::new(prefix.as_os_str()));
-                }
-                Component::ParentDir => relative.push(Path::new("..")),
-                Component::Normal(part) => relative.push(Path::new(part)),
-            }
-        }
-
-        if relative.as_os_str().is_empty() {
-            None
-        } else {
-            Some(relative)
-        }
-    }
-
-    /// Returns the source path.
-    #[must_use]
-    pub fn path(&self) -> &Path {
-        &self.path
-    }
-
-    /// Reports whether the directory contents should be copied.
-    #[must_use]
-    pub const fn copy_contents(&self) -> bool {
-        self.copy_contents
-    }
-}
-
-fn detect_relative_prefix_components(operand: &OsStr) -> Option<usize> {
-    use std::path::{Component, Path};
-
-    let path = Path::new(operand);
-
-    #[cfg(unix)]
-    if let Some(count) = detect_marker_components_unix(operand) {
-        return Some(count);
-    }
-
-    #[cfg(windows)]
-    if let Some(count) = detect_marker_components_windows(operand) {
-        return Some(count);
-    }
-
-    let components: Vec<Component<'_>> = path.components().collect();
-
-    if components.is_empty() {
-        return None;
-    }
-
-    let mut skip = 0;
-
-    if let Some(Component::Prefix(_)) = components.first() {
-        if !path.has_root() {
-            return None;
-        }
-        skip += 1;
-    }
-
-    if let Some(Component::RootDir) = components.get(skip) {
-        skip += 1;
-    }
-
-    if skip > 0 { Some(skip) } else { None }
-}
-
-#[derive(Clone, Copy, Debug, Default, Eq, PartialEq)]
-struct DestinationState {
-    exists: bool,
-    is_dir: bool,
-    symlink_to_dir: bool,
-}
-
-#[derive(Debug)]
-struct DirectoryEntry {
-    file_name: OsString,
-    path: PathBuf,
-    metadata: fs::Metadata,
-}
-
-/// Destination operand capturing directory semantics requested by the caller.
-#[derive(Clone, Debug, Eq, PartialEq)]
-struct DestinationSpec {
-    path: PathBuf,
-    force_directory: bool,
-}
-
-impl DestinationSpec {
-    fn from_operand(operand: &OsString) -> Self {
-        let force_directory = has_trailing_separator(operand.as_os_str());
-        Self {
-            path: PathBuf::from(operand),
-            force_directory,
-        }
-    }
-
-    /// Returns the destination path supplied by the caller.
-    #[must_use]
-    pub fn path(&self) -> &Path {
-        &self.path
-    }
-
-    /// Reports whether the operand explicitly requested directory semantics.
-    #[must_use]
-    pub const fn force_directory(&self) -> bool {
-        self.force_directory
-    }
-}
-
-#[cfg(unix)]
-fn detect_marker_components_unix(operand: &OsStr) -> Option<usize> {
-    use std::os::unix::ffi::OsStrExt;
-
-    let bytes = operand.as_bytes();
-    if bytes.is_empty() {
-        return None;
-    }
-
-    let mut index = 0;
-    let len = bytes.len();
-    let mut component_count = 0;
-
-    if bytes[0] == b'/' {
-        component_count += 1;
-        while index < len && bytes[index] == b'/' {
-            index += 1;
-        }
-    }
-
-    if index >= len {
-        return None;
-    }
-
-    let mut start = index;
-    let mut count = component_count;
-
-    while index <= len {
-        if index == len || bytes[index] == b'/' {
-            if start != index {
-                let component = &bytes[start..index];
-                if component == b"." {
-                    return Some(count);
-                }
-                count += 1;
-            }
-
-            while index < len && bytes[index] == b'/' {
-                index += 1;
-            }
-            start = index;
-            if index == len {
-                break;
-            }
-        } else {
-            index += 1;
-        }
-    }
-
-    None
-}
-
-#[cfg(windows)]
-fn detect_marker_components_windows(operand: &OsStr) -> Option<usize> {
-    use std::os::windows::ffi::OsStrExt;
-
-    fn is_separator(unit: u16) -> bool {
-        unit == b'/' as u16 || unit == b'\\' as u16
-    }
-
-    fn is_single_dot(units: &[u16]) -> bool {
-        units.len() == 1 && units[0] == b'.' as u16
-    }
-
-    let units: Vec<u16> = operand.encode_wide().collect();
-    if units.is_empty() {
-        return None;
-    }
-
-    let len = units.len();
-    let mut index = 0;
-    let mut count = 0;
-
-    if len >= 2 && units[1] == b':' as u16 {
-        count += 1;
-        index = 2;
-        if index < len && is_separator(units[index]) {
-            count += 1;
-            while index < len && is_separator(units[index]) {
-                index += 1;
-            }
-        }
-    } else if len >= 2 && is_separator(units[0]) && is_separator(units[1]) {
-        count += 1;
-        index = 2;
-        while index < len && !is_separator(units[index]) {
-            index += 1;
-        }
-        if index < len && is_separator(units[index]) {
-            index += 1;
-        }
-        while index < len && !is_separator(units[index]) {
-            index += 1;
-        }
-        if index < len && is_separator(units[index]) {
-            count += 1;
-            while index < len && is_separator(units[index]) {
-                index += 1;
-            }
-        }
-    } else if is_separator(units[0]) {
-        count += 1;
-        while index < len && is_separator(units[index]) {
-            index += 1;
-        }
-    }
-
-    if index >= len {
-        return None;
-    }
-
-    let mut start = index;
-    let mut components = count;
-
-    while index <= len {
-        if index == len || is_separator(units[index]) {
-            if start != index {
-                let component = &units[start..index];
-                if is_single_dot(component) {
-                    return Some(components);
-                }
-                components += 1;
-            }
-
-            while index < len && is_separator(units[index]) {
-                index += 1;
-            }
-            start = index;
-            if index == len {
-                break;
-            }
-        } else {
-            index += 1;
-        }
-    }
-
-    None
-}
-
-fn non_empty_path(path: &Path) -> Option<&Path> {
-    if path.as_os_str().is_empty() {
-        None
-    } else {
-        Some(path)
-    }
-}
-
-fn follow_symlink_metadata(path: &Path) -> Result<fs::Metadata, LocalCopyError> {
-    fs::metadata(path)
-        .map_err(|error| LocalCopyError::io("inspect symlink target", path.to_path_buf(), error))
-}
-
-fn copy_sources(
-    plan: &LocalCopyPlan,
-    mode: LocalCopyExecution,
-    options: LocalCopyOptions,
-    handler: Option<&mut dyn LocalCopyRecordHandler>,
-) -> Result<CopyOutcome, LocalCopyError> {
-    let destination_root = plan.destination_spec().path().to_path_buf();
-    let mut context = CopyContext::new(mode, options, handler, destination_root);
-    let result = {
-        let context = &mut context;
-        (|| -> Result<(), LocalCopyError> {
-            let multiple_sources = plan.sources().len() > 1;
-            let destination_path = plan.destination_spec().path();
-            let mut destination_state = query_destination_state(destination_path)?;
-            if context.keep_dirlinks_enabled() && destination_state.symlink_to_dir {
-                destination_state.is_dir = true;
-            }
-
-            if plan.destination_spec().force_directory() {
-                ensure_destination_directory(
-                    destination_path,
-                    &mut destination_state,
-                    context.mode(),
-                )?;
-            }
-
-            if multiple_sources {
-                ensure_destination_directory(
-                    destination_path,
-                    &mut destination_state,
-                    context.mode(),
-                )?;
-            }
-
-            let destination_behaves_like_directory =
-                destination_state.is_dir || plan.destination_spec().force_directory();
-
-            let relative_enabled = context.relative_paths_enabled();
-
-            for source in plan.sources() {
-                context.enforce_timeout()?;
-                let source_path = source.path();
-                let metadata_start = Instant::now();
-                let metadata = match fs::symlink_metadata(source_path) {
-                    Ok(metadata) => metadata,
-                    Err(error)
-                        if error.kind() == io::ErrorKind::NotFound
-                            && context.ignore_missing_args_enabled() =>
-                    {
-                        context.record_file_list_generation(metadata_start.elapsed());
-                        continue;
-                    }
-                    Err(error) => {
-                        return Err(LocalCopyError::io(
-                            "access source",
-                            source_path.to_path_buf(),
-                            error,
-                        ));
-                    }
-                };
-                context.record_file_list_generation(metadata_start.elapsed());
-                let file_type = metadata.file_type();
-                let metadata_options = context.metadata_options();
-
-                let root_device = if context.one_file_system_enabled() {
-                    device_identifier(source_path, &metadata)
-                } else {
-                    None
-                };
-
-                let relative_root = if relative_enabled {
-                    source.relative_root()
-                } else {
-                    None
-                };
-                let relative_root = relative_root.filter(|path| !path.as_os_str().is_empty());
-                let relative_parent = relative_root
-                    .as_ref()
-                    .and_then(|root| root.parent().map(|parent| parent.to_path_buf()))
-                    .filter(|parent| !parent.as_os_str().is_empty());
-
-                let requires_directory_destination = relative_parent.is_some()
-                    || (relative_root.is_some() && (source.copy_contents() || file_type.is_dir()));
-
-                if requires_directory_destination && !destination_behaves_like_directory {
-                    return Err(LocalCopyError::invalid_argument(
-                        LocalCopyArgumentError::DestinationMustBeDirectory,
-                    ));
-                }
-
-                let destination_base = if let Some(parent) = &relative_parent {
-                    destination_path.join(parent)
-                } else {
-                    destination_path.to_path_buf()
-                };
-
-                if file_type.is_dir() {
-                    if source.copy_contents() {
-                        if let Some(root) = relative_root.as_ref() {
-                            if !context.allows(root.as_path(), true) {
-                                continue;
-                            }
-                        }
-
-                        let mut target_root = destination_path.to_path_buf();
-                        if let Some(root) = &relative_root {
-                            target_root = destination_path.join(root);
-                        }
-
-                        copy_directory_recursive(
-                            context,
-                            source_path,
-                            &target_root,
-                            &metadata,
-                            relative_root
-                                .as_ref()
-                                .and_then(|root| non_empty_path(root.as_path())),
-                            root_device,
-                        )?;
-                        continue;
-                    }
-
-                    let name = source_path.file_name().ok_or_else(|| {
-                        LocalCopyError::invalid_argument(
-                            LocalCopyArgumentError::DirectoryNameUnavailable,
-                        )
-                    })?;
-                    let relative = relative_root
-                        .clone()
-                        .unwrap_or_else(|| PathBuf::from(Path::new(name)));
-                    if !context.allows(&relative, true) {
-                        continue;
-                    }
-
-                    let target = if destination_behaves_like_directory || multiple_sources {
-                        destination_base.join(name)
-                    } else {
-                        destination_path.to_path_buf()
-                    };
-
-                    copy_directory_recursive(
-                        context,
-                        source_path,
-                        &target,
-                        &metadata,
-                        non_empty_path(relative.as_path()),
-                        root_device,
-                    )?;
-                } else {
-                    let name = source_path.file_name().ok_or_else(|| {
-                        LocalCopyError::invalid_argument(
-                            LocalCopyArgumentError::FileNameUnavailable,
-                        )
-                    })?;
-                    let relative = relative_root
-                        .clone()
-                        .unwrap_or_else(|| PathBuf::from(Path::new(name)));
-                    let followed_metadata = if file_type.is_symlink()
-                        && (context.copy_links_enabled() || context.copy_dirlinks_enabled())
-                    {
-                        match follow_symlink_metadata(source_path) {
-                            Ok(target_metadata) => Some(target_metadata),
-                            Err(error) => {
-                                if context.copy_links_enabled() {
-                                    return Err(error);
-                                }
-                                None
-                            }
-                        }
-                    } else {
-                        None
-                    };
-
-                    let (effective_metadata, effective_type) =
-                        if let Some(ref target_metadata) = followed_metadata {
-                            let ty = target_metadata.file_type();
-                            if context.copy_links_enabled()
-                                || (context.copy_dirlinks_enabled() && ty.is_dir())
-                            {
-                                (target_metadata, ty)
-                            } else {
-                                (&metadata, file_type)
-                            }
-                        } else {
-                            (&metadata, file_type)
-                        };
-
-                    if !context.allows(&relative, effective_type.is_dir()) {
-                        continue;
-                    }
-
-                    let target = if destination_behaves_like_directory {
-                        destination_base.join(name)
-                    } else {
-                        destination_path.to_path_buf()
-                    };
-
-                    let record_path = non_empty_path(relative.as_path());
-                    if effective_type.is_file() {
-                        copy_file(
-                            context,
-                            source_path,
-                            &target,
-                            effective_metadata,
-                            record_path,
-                        )?;
-                    } else if effective_type.is_dir() {
-                        copy_directory_recursive(
-                            context,
-                            source_path,
-                            &target,
-                            effective_metadata,
-                            non_empty_path(relative.as_path()),
-                            root_device,
-                        )?;
-                    } else if file_type.is_symlink() && !context.copy_links_enabled() {
-                        copy_symlink(
-                            context,
-                            source_path,
-                            &target,
-                            &metadata,
-                            &metadata_options,
-                            record_path,
-                        )?;
-                    } else if is_fifo(&effective_type) {
-                        if !context.specials_enabled() {
-                            context.record_skipped_non_regular(record_path);
-                            continue;
-                        }
-                        copy_fifo(
-                            context,
-                            source_path,
-                            &target,
-                            effective_metadata,
-                            &metadata_options,
-                            record_path,
-                        )?;
-                    } else if is_device(&effective_type) {
-                        if !context.devices_enabled() {
-                            context.record_skipped_non_regular(record_path);
-                            continue;
-                        }
-                        copy_device(
-                            context,
-                            source_path,
-                            &target,
-                            effective_metadata,
-                            &metadata_options,
-                            record_path,
-                        )?;
-                    } else {
-                        return Err(LocalCopyError::invalid_argument(
-                            LocalCopyArgumentError::UnsupportedFileType,
-                        ));
-                    }
-                }
-
-                context.enforce_timeout()?;
-            }
-
-            context.flush_deferred_updates()?;
-            context.flush_deferred_deletions()?;
-            context.enforce_timeout()?;
-            Ok(())
-        })()
-    };
-
-    match result {
-        Ok(()) => Ok(context.into_outcome()),
-        Err(error) => {
-            context.rollback_on_error(&error);
-            Err(error)
-        }
-    }
-}
-fn query_destination_state(path: &Path) -> Result<DestinationState, LocalCopyError> {
-    match fs::symlink_metadata(path) {
-        Ok(metadata) => {
-            let file_type = metadata.file_type();
-            let symlink_to_dir = if file_type.is_symlink() {
-                follow_symlink_metadata(path)
-                    .map(|target| target.file_type().is_dir())
-                    .unwrap_or(false)
-            } else {
-                false
-            };
-
-            Ok(DestinationState {
-                exists: true,
-                is_dir: file_type.is_dir(),
-                symlink_to_dir,
-            })
-        }
-        Err(error) if error.kind() == io::ErrorKind::NotFound => Ok(DestinationState::default()),
-        Err(error) => Err(LocalCopyError::io(
-            "inspect destination",
-            path.to_path_buf(),
-            error,
-        )),
-    }
-}
-
-fn copy_directory_recursive(
-    context: &mut CopyContext,
-    source: &Path,
-    destination: &Path,
-    metadata: &fs::Metadata,
-    relative: Option<&Path>,
-    root_device: Option<u64>,
-) -> Result<bool, LocalCopyError> {
-    #[cfg(any(feature = "acl", feature = "xattr"))]
-    let mode = context.mode();
-    #[cfg(not(any(feature = "acl", feature = "xattr")))]
-    let _mode = context.mode();
-    #[cfg(feature = "xattr")]
-    let preserve_xattrs = context.xattrs_enabled();
-    #[cfg(feature = "acl")]
-    let preserve_acls = context.acls_enabled();
-    let prune_enabled = context.prune_empty_dirs_enabled();
-
-    let root_device = if context.one_file_system_enabled() {
-        root_device.or_else(|| device_identifier(source, metadata))
-    } else {
-        None
-    };
-
-    let mut destination_missing = false;
-
-    let keep_dirlinks = context.keep_dirlinks_enabled();
-
-    match fs::symlink_metadata(destination) {
-        Ok(existing) => {
-            let file_type = existing.file_type();
-            if file_type.is_dir() {
-                // Directory already present; nothing to do.
-            } else if file_type.is_symlink() && keep_dirlinks {
-                let target_metadata = follow_symlink_metadata(destination)?;
-                if !target_metadata.file_type().is_dir() {
-                    return Err(LocalCopyError::invalid_argument(
-                        LocalCopyArgumentError::ReplaceNonDirectoryWithDirectory,
-                    ));
-                }
-            } else {
-                return Err(LocalCopyError::invalid_argument(
-                    LocalCopyArgumentError::ReplaceNonDirectoryWithDirectory,
-                ));
-            }
-        }
-        Err(error) if error.kind() == io::ErrorKind::NotFound => {
-            destination_missing = true;
-        }
-        Err(error) => {
-            return Err(LocalCopyError::io(
-                "inspect destination directory",
-                destination.to_path_buf(),
-                error,
-            ));
-        }
-    }
-
-    let list_start = Instant::now();
-    let entries = read_directory_entries_sorted(source)?;
-    context.record_file_list_generation(list_start.elapsed());
-    context.register_progress();
-
-    let dir_merge_guard = context.enter_directory(source)?;
-    if dir_merge_guard.is_excluded() {
-        return Ok(false);
-    }
-    let _dir_merge_guard = dir_merge_guard;
-
-    let directory_ready = Cell::new(!destination_missing);
-    let mut created_directory_on_disk = false;
-    let creation_record_pending = destination_missing && relative.is_some();
-    let mut pending_record: Option<LocalCopyRecord> = None;
-    let metadata_record = relative.map(|rel| {
-        (
-            rel.to_path_buf(),
-            LocalCopyMetadata::from_metadata(metadata, None),
-        )
-    });
-
-    let mut kept_any = !prune_enabled;
-
-    let mut ensure_directory = |context: &mut CopyContext| -> Result<(), LocalCopyError> {
-        if directory_ready.get() {
-            return Ok(());
-        }
-
-        if context.mode().is_dry_run() {
-            if !context.implied_dirs_enabled() {
-                if let Some(parent) = destination.parent() {
-                    context.prepare_parent_directory(parent)?;
-                }
-            }
-            directory_ready.set(true);
-        } else {
-            if let Some(parent) = destination.parent() {
-                context.prepare_parent_directory(parent)?;
-            }
-            if context.implied_dirs_enabled() {
-                fs::create_dir_all(destination).map_err(|error| {
-                    LocalCopyError::io("create directory", destination.to_path_buf(), error)
-                })?;
-            } else {
-                fs::create_dir(destination).map_err(|error| {
-                    LocalCopyError::io("create directory", destination.to_path_buf(), error)
-                })?;
-            }
-            context.register_progress();
-            context.register_created_path(destination, CreatedEntryKind::Directory, false);
-            directory_ready.set(true);
-            created_directory_on_disk = true;
-        }
-
-        if pending_record.is_none() {
-            if let Some((ref rel_path, ref snapshot)) = metadata_record {
-                pending_record = Some(LocalCopyRecord::new(
-                    rel_path.clone(),
-                    LocalCopyAction::DirectoryCreated,
-                    0,
-                    Some(snapshot.len()),
-                    Duration::default(),
-                    Some(snapshot.clone()),
-                ));
-            }
-        }
-
-        Ok(())
-    };
-
-    if !directory_ready.get() && !prune_enabled {
-        ensure_directory(context)?;
-    }
-
-    #[derive(Clone, Copy)]
-    enum EntryAction {
-        SkipExcluded,
-        SkipNonRegular,
-        SkipMountPoint,
-        CopyDirectory,
-        CopyFile,
-        CopySymlink,
-        CopyFifo,
-        CopyDevice,
-    }
-
-    struct PlannedEntry<'a> {
-        entry: &'a DirectoryEntry,
-        relative: PathBuf,
-        action: EntryAction,
-        metadata_override: Option<fs::Metadata>,
-    }
-
-    impl<'a> PlannedEntry<'a> {
-        fn metadata(&self) -> &fs::Metadata {
-            self.metadata_override
-                .as_ref()
-                .unwrap_or(&self.entry.metadata)
-        }
-    }
-
-    let deletion_enabled = context.options().delete_extraneous();
-    let delete_timing = context.delete_timing();
-    let mut keep_names = if deletion_enabled {
-        Vec::with_capacity(entries.len())
-    } else {
-        Vec::new()
-    };
-    let mut planned_entries = Vec::with_capacity(entries.len());
-
-    for entry in entries.iter() {
-        context.enforce_timeout()?;
-        context.register_progress();
-
-        let file_name = entry.file_name.clone();
-        let entry_metadata = &entry.metadata;
-        let entry_type = entry_metadata.file_type();
-        let mut metadata_override = None;
-        let mut effective_type = entry_type;
-        if entry_type.is_symlink()
-            && (context.copy_links_enabled() || context.copy_dirlinks_enabled())
-        {
-            match follow_symlink_metadata(entry.path.as_path()) {
-                Ok(target_metadata) => {
-                    let target_type = target_metadata.file_type();
-                    if context.copy_links_enabled()
-                        || (context.copy_dirlinks_enabled() && target_type.is_dir())
-                    {
-                        effective_type = target_type;
-                        metadata_override = Some(target_metadata);
-                    }
-                }
-                Err(error) => {
-                    if context.copy_links_enabled() {
-                        return Err(error);
-                    }
-                }
-            }
-        }
-        let relative_path = match relative {
-            Some(base) => base.join(Path::new(&file_name)),
-            None => PathBuf::from(Path::new(&file_name)),
-        };
-
-        let mut keep_name = true;
-
-        let mut action = if !context.allows(&relative_path, effective_type.is_dir()) {
-            // Skip excluded entries while optionally allowing deletion sweeps to remove them.
-            if context.options().delete_excluded_enabled() {
-                keep_name = false;
-            }
-            EntryAction::SkipExcluded
-        } else if entry_type.is_dir() {
-            EntryAction::CopyDirectory
-        } else if effective_type.is_file() {
-            EntryAction::CopyFile
-        } else if effective_type.is_dir() {
-            EntryAction::CopyDirectory
-        } else if entry_type.is_symlink() && !context.copy_links_enabled() {
-            EntryAction::CopySymlink
-        } else if is_fifo(&effective_type) {
-            if context.specials_enabled() {
-                EntryAction::CopyFifo
-            } else {
-                keep_name = false;
-                EntryAction::SkipNonRegular
-            }
-        } else if is_device(&effective_type) {
-            if context.devices_enabled() {
-                EntryAction::CopyDevice
-            } else {
-                keep_name = false;
-                EntryAction::SkipNonRegular
-            }
-        } else {
-            return Err(LocalCopyError::invalid_argument(
-                LocalCopyArgumentError::UnsupportedFileType,
-            ));
-        };
-
-        if matches!(action, EntryAction::CopySymlink)
-            && context.safe_links_enabled()
-            && context.copy_unsafe_links_enabled()
-        {
-            match fs::read_link(entry.path.as_path()) {
-                Ok(target) => {
-                    if !symlink_target_is_safe(&target, relative_path.as_path()) {
-                        match follow_symlink_metadata(entry.path.as_path()) {
-                            Ok(target_metadata) => {
-                                let target_type = target_metadata.file_type();
-                                if target_type.is_dir() {
-                                    action = EntryAction::CopyDirectory;
-                                    metadata_override = Some(target_metadata);
-                                } else if target_type.is_file() {
-                                    action = EntryAction::CopyFile;
-                                    metadata_override = Some(target_metadata);
-                                } else if is_fifo(&target_type) {
-                                    if context.specials_enabled() {
-                                        action = EntryAction::CopyFifo;
-                                        metadata_override = Some(target_metadata);
-                                    } else {
-                                        keep_name = false;
-                                        action = EntryAction::SkipNonRegular;
-                                        metadata_override = None;
-                                    }
-                                } else if is_device(&target_type) {
-                                    if context.devices_enabled() {
-                                        action = EntryAction::CopyDevice;
-                                        metadata_override = Some(target_metadata);
-                                    } else {
-                                        keep_name = false;
-                                        action = EntryAction::SkipNonRegular;
-                                        metadata_override = None;
-                                    }
-                                } else {
-                                    return Err(LocalCopyError::invalid_argument(
-                                        LocalCopyArgumentError::UnsupportedFileType,
-                                    ));
-                                }
-                            }
-                            Err(error) => {
-                                return Err(error);
-                            }
-                        }
-                    }
-                }
-                Err(error) => {
-                    return Err(LocalCopyError::io(
-                        "read symbolic link",
-                        entry.path.to_path_buf(),
-                        error,
-                    ));
-                }
-            }
-        }
-
-        if matches!(action, EntryAction::CopyDirectory) && context.one_file_system_enabled() {
-            if let Some(root) = root_device {
-                if let Some(entry_device) = device_identifier(
-                    entry.path.as_path(),
-                    metadata_override.as_ref().unwrap_or(entry_metadata),
-                ) {
-                    if entry_device != root {
-                        action = EntryAction::SkipMountPoint;
-                    }
-                }
-            }
-        }
-
-        if deletion_enabled && keep_name {
-            let preserve_name = match delete_timing {
-                Some(DeleteTiming::Before) => matches!(
-                    action,
-                    EntryAction::CopyDirectory
-                        | EntryAction::SkipExcluded
-                        | EntryAction::SkipMountPoint
-                ),
-                _ => true,
-            };
-
-            if preserve_name {
-                keep_names.push(file_name.clone());
-            }
-        }
-
-        planned_entries.push(PlannedEntry {
-            entry,
-            relative: relative_path,
-            action,
-            metadata_override,
-        });
-    }
-
-    if deletion_enabled && matches!(delete_timing, Some(DeleteTiming::Before)) {
-        delete_extraneous_entries(context, destination, relative, &keep_names)?;
-    }
-
-    for planned in planned_entries {
-        let file_name = &planned.entry.file_name;
-        let target_path = destination.join(Path::new(file_name));
-        let entry_metadata = planned.metadata();
-        let record_relative = non_empty_path(planned.relative.as_path());
-
-        match planned.action {
-            EntryAction::SkipExcluded => {}
-            EntryAction::SkipNonRegular => {
-                context.record_skipped_non_regular(record_relative);
-            }
-            EntryAction::SkipMountPoint => {
-                context.record_skipped_mount_point(record_relative);
-            }
-            EntryAction::CopyDirectory => {
-                ensure_directory(context)?;
-                let child_kept = copy_directory_recursive(
-                    context,
-                    planned.entry.path.as_path(),
-                    &target_path,
-                    entry_metadata,
-                    Some(planned.relative.as_path()),
-                    root_device,
-                )?;
-                if child_kept {
-                    kept_any = true;
-                }
-            }
-            EntryAction::CopyFile => {
-                ensure_directory(context)?;
-                copy_file(
-                    context,
-                    planned.entry.path.as_path(),
-                    &target_path,
-                    entry_metadata,
-                    Some(planned.relative.as_path()),
-                )?;
-                kept_any = true;
-            }
-            EntryAction::CopySymlink => {
-                ensure_directory(context)?;
-                let metadata_options = context.metadata_options();
-                copy_symlink(
-                    context,
-                    planned.entry.path.as_path(),
-                    &target_path,
-                    entry_metadata,
-                    &metadata_options,
-                    Some(planned.relative.as_path()),
-                )?;
-                kept_any = true;
-            }
-            EntryAction::CopyFifo => {
-                ensure_directory(context)?;
-                let metadata_options = context.metadata_options();
-                copy_fifo(
-                    context,
-                    planned.entry.path.as_path(),
-                    &target_path,
-                    entry_metadata,
-                    &metadata_options,
-                    Some(planned.relative.as_path()),
-                )?;
-                kept_any = true;
-            }
-            EntryAction::CopyDevice => {
-                ensure_directory(context)?;
-                let metadata_options = context.metadata_options();
-                copy_device(
-                    context,
-                    planned.entry.path.as_path(),
-                    &target_path,
-                    entry_metadata,
-                    &metadata_options,
-                    Some(planned.relative.as_path()),
-                )?;
-                kept_any = true;
-            }
-        }
-    }
-
-    if deletion_enabled {
-        match delete_timing.unwrap_or(DeleteTiming::During) {
-            DeleteTiming::Before => {}
-            DeleteTiming::During => {
-                delete_extraneous_entries(context, destination, relative, &keep_names)?;
-            }
-            DeleteTiming::Delay | DeleteTiming::After => {
-                let relative_owned = relative.map(Path::to_path_buf);
-                context.defer_deletion(destination.to_path_buf(), relative_owned, keep_names);
-            }
-        }
-    }
-
-    if prune_enabled && !kept_any {
-        if created_directory_on_disk {
-            fs::remove_dir(destination).map_err(|error| {
-                LocalCopyError::io("remove empty directory", destination.to_path_buf(), error)
-            })?;
-            if context
-                .last_created_entry_path()
-                .is_some_and(|path| path == destination)
-            {
-                context.pop_last_created_entry();
-            }
-        }
-        return Ok(false);
-    }
-
-    context.summary_mut().record_directory_total();
-    if creation_record_pending {
-        context.summary_mut().record_directory();
-    }
-    if let Some(record) = pending_record {
-        context.record(record);
-    }
-
-    if !context.mode().is_dry_run() {
-        let metadata_options = if context.omit_dir_times_enabled() {
-            context.metadata_options().preserve_times(false)
-        } else {
-            context.metadata_options()
-        };
-        apply_directory_metadata_with_options(destination, metadata, metadata_options)
-            .map_err(map_metadata_error)?;
-        #[cfg(feature = "xattr")]
-        sync_xattrs_if_requested(preserve_xattrs, mode, source, destination, true)?;
-        #[cfg(feature = "acl")]
-        sync_acls_if_requested(preserve_acls, mode, source, destination, true)?;
-    }
-
-    Ok(true)
-}
-
-#[derive(Debug)]
-enum ReferenceDecision {
-    Skip,
-    Copy(PathBuf),
-    Link(PathBuf),
-}
-
-fn resolve_reference_candidate(base: &Path, relative: &Path, destination: &Path) -> PathBuf {
-    if base.is_absolute() {
-        base.join(relative)
-    } else {
-        let mut ancestor = destination.to_path_buf();
-        let depth = relative.components().count();
-        for _ in 0..depth {
-            if !ancestor.pop() {
-                break;
-            }
-        }
-        ancestor.join(base).join(relative)
-    }
-}
-
-struct ReferenceQuery<'a> {
-    destination: &'a Path,
-    relative: &'a Path,
-    source: &'a Path,
-    metadata: &'a fs::Metadata,
-    metadata_options: &'a MetadataOptions,
-    size_only: bool,
-    checksum: bool,
-}
-
-fn find_reference_action(
-    context: &CopyContext<'_>,
-    query: ReferenceQuery<'_>,
-) -> Result<Option<ReferenceDecision>, LocalCopyError> {
-    let ReferenceQuery {
-        destination,
-        relative,
-        source,
-        metadata,
-        metadata_options,
-        size_only,
-        checksum,
-    } = query;
-    for reference in context.reference_directories() {
-        let candidate = resolve_reference_candidate(reference.path(), relative, destination);
-        let candidate_metadata = match fs::symlink_metadata(&candidate) {
-            Ok(meta) => meta,
-            Err(error) if error.kind() == io::ErrorKind::NotFound => continue,
-            Err(error) => {
-                return Err(LocalCopyError::io(
-                    "inspect reference file",
-                    candidate,
-                    error,
-                ));
-            }
-        };
-
-        if !candidate_metadata.file_type().is_file() {
-            continue;
-        }
-
-        if should_skip_copy(CopyComparison {
-            source_path: source,
-            source: metadata,
-            destination_path: &candidate,
-            destination: &candidate_metadata,
-            options: metadata_options,
-            size_only,
-            checksum,
-            checksum_algorithm: context.options().checksum_algorithm(),
-            modify_window: context.options().modify_window(),
-        }) {
-            return Ok(Some(match reference.kind() {
-                ReferenceDirectoryKind::Compare => ReferenceDecision::Skip,
-                ReferenceDirectoryKind::Copy => ReferenceDecision::Copy(candidate),
-                ReferenceDirectoryKind::Link => ReferenceDecision::Link(candidate),
-            }));
-        }
-    }
-
-    Ok(None)
-}
-
-fn maybe_preallocate_destination(
-    file: &mut fs::File,
-    path: &Path,
-    total_len: u64,
-    existing_bytes: u64,
-    enabled: bool,
-) -> Result<(), LocalCopyError> {
-    if !enabled || total_len == 0 || total_len <= existing_bytes {
-        return Ok(());
-    }
-
-    preallocate_destination_file(file, path, total_len)
-}
-
-fn preallocate_destination_file(
-    file: &mut fs::File,
-    path: &Path,
-    total_len: u64,
-) -> Result<(), LocalCopyError> {
-    #[cfg(unix)]
-    {
-        if total_len == 0 {
-            return Ok(());
-        }
-
-        if total_len > i64::MAX as u64 {
-            return Err(LocalCopyError::io(
-                "preallocate destination file",
-                path.to_path_buf(),
-                io::Error::new(
-                    io::ErrorKind::InvalidInput,
-                    "preallocation size exceeds platform limit",
-                ),
-            ));
-        }
-
-        let fd = file.as_fd();
-        match fallocate(fd, FallocateFlags::empty(), 0, total_len) {
-            Ok(()) => Ok(()),
-            Err(Errno::OPNOTSUPP | Errno::NOSYS | Errno::INVAL) => {
-                file.set_len(total_len).map_err(|error| {
-                    LocalCopyError::io("preallocate destination file", path.to_path_buf(), error)
-                })
-            }
-            Err(errno) => Err(LocalCopyError::io(
-                "preallocate destination file",
-                path.to_path_buf(),
-                io::Error::from_raw_os_error(errno.raw_os_error()),
-            )),
-        }
-    }
-
-    #[cfg(not(unix))]
-    {
-        if total_len == 0 {
-            return Ok(());
-        }
-
-        file.set_len(total_len).map_err(|error| {
-            LocalCopyError::io("preallocate destination file", path.to_path_buf(), error)
-        })
-    }
-}
-
-fn copy_file(
-    context: &mut CopyContext,
-    source: &Path,
-    destination: &Path,
-    metadata: &fs::Metadata,
-    relative: Option<&Path>,
-) -> Result<(), LocalCopyError> {
-    context.enforce_timeout()?;
-    let metadata_options = context.metadata_options();
-    let mode = context.mode();
-    let file_type = metadata.file_type();
-    #[cfg(feature = "xattr")]
-    let preserve_xattrs = context.xattrs_enabled();
-    #[cfg(feature = "acl")]
-    let preserve_acls = context.acls_enabled();
-    let record_path = relative
-        .map(Path::to_path_buf)
-        .or_else(|| source.file_name().map(PathBuf::from))
-        .unwrap_or_else(|| {
-            destination
-                .file_name()
-                .map(PathBuf::from)
-                .unwrap_or_default()
-        });
-    let file_size = metadata.len();
-    context.summary_mut().record_regular_file_total();
-    context.summary_mut().record_total_bytes(file_size);
-
-    if let Some(min_limit) = context.min_file_size_limit() {
-        if file_size < min_limit {
-            return Ok(());
-        }
-    }
-
-    if let Some(max_limit) = context.max_file_size_limit() {
-        if file_size > max_limit {
-            return Ok(());
-        }
-    }
-    if let Some(parent) = destination.parent() {
-        context.prepare_parent_directory(parent)?;
-    }
-
-    let existing_metadata = match fs::symlink_metadata(destination) {
-        Ok(existing) => Some(existing),
-        Err(error) if error.kind() == io::ErrorKind::NotFound => None,
-        Err(error) => {
-            return Err(LocalCopyError::io(
-                "inspect existing destination",
-                destination.to_path_buf(),
-                error,
-            ));
-        }
-    };
-
-    if let Some(existing) = &existing_metadata {
-        if existing.file_type().is_dir() {
-            return Err(LocalCopyError::invalid_argument(
-                LocalCopyArgumentError::ReplaceDirectoryWithFile,
-            ));
-        }
-    }
-
-    let destination_previously_existed = existing_metadata.is_some();
-
-    if mode.is_dry_run() {
-        if context.update_enabled() {
-            if let Some(existing) = existing_metadata.as_ref() {
-                if destination_is_newer(metadata, existing) {
-                    context.summary_mut().record_regular_file_skipped_newer();
-                    let metadata_snapshot = LocalCopyMetadata::from_metadata(metadata, None);
-                    let total_bytes = Some(metadata_snapshot.len());
-                    context.record(LocalCopyRecord::new(
-                        record_path.clone(),
-                        LocalCopyAction::SkippedNewerDestination,
-                        0,
-                        total_bytes,
-                        Duration::default(),
-                        Some(metadata_snapshot),
-                    ));
-                    return Ok(());
-                }
-            }
-        }
-
-        if context.ignore_existing_enabled() && existing_metadata.is_some() {
-            context.summary_mut().record_regular_file_ignored_existing();
-            let metadata_snapshot = LocalCopyMetadata::from_metadata(metadata, None);
-            let total_bytes = Some(metadata_snapshot.len());
-            context.record(LocalCopyRecord::new(
-                record_path.clone(),
-                LocalCopyAction::SkippedExisting,
-                0,
-                total_bytes,
-                Duration::default(),
-                Some(metadata_snapshot),
-            ));
-            return Ok(());
-        }
-
-        let mut reader = fs::File::open(source)
-            .map_err(|error| LocalCopyError::io("open source file", source.to_path_buf(), error))?;
-
-        let append_mode = determine_append_mode(
-            context.append_enabled(),
-            context.append_verify_enabled(),
-            &mut reader,
-            source,
-            destination,
-            existing_metadata.as_ref(),
-            file_size,
-        )?;
-        let append_offset = match append_mode {
-            AppendMode::Append(offset) => offset,
-            AppendMode::Disabled => 0,
-        };
-        let bytes_transferred = file_size.saturating_sub(append_offset);
-
-        context
-            .summary_mut()
-            .record_file(file_size, bytes_transferred, None);
-        let metadata_snapshot = LocalCopyMetadata::from_metadata(metadata, None);
-        let total_bytes = Some(metadata_snapshot.len());
-        context.record(
-            LocalCopyRecord::new(
-                record_path.clone(),
-                LocalCopyAction::DataCopied,
-                bytes_transferred,
-                total_bytes,
-                Duration::default(),
-                Some(metadata_snapshot),
-            )
-            .with_creation(!destination_previously_existed),
-        );
-        remove_source_entry_if_requested(context, source, Some(record_path.as_path()), file_type)?;
-        return Ok(());
-    }
-
-    if context.update_enabled() {
-        if let Some(existing) = existing_metadata.as_ref() {
-            if destination_is_newer(metadata, existing) {
-                context.summary_mut().record_regular_file_skipped_newer();
-                context.record_hard_link(metadata, destination);
-                let metadata_snapshot = LocalCopyMetadata::from_metadata(metadata, None);
-                let total_bytes = Some(metadata_snapshot.len());
-                context.record(LocalCopyRecord::new(
-                    record_path.clone(),
-                    LocalCopyAction::SkippedNewerDestination,
-                    0,
-                    total_bytes,
-                    Duration::default(),
-                    Some(metadata_snapshot),
-                ));
-                return Ok(());
-            }
-        }
-    }
-
-    if context.ignore_existing_enabled() && existing_metadata.is_some() {
-        context.summary_mut().record_regular_file_ignored_existing();
-        context.record_hard_link(metadata, destination);
-        let metadata_snapshot = LocalCopyMetadata::from_metadata(metadata, None);
-        let total_bytes = Some(metadata_snapshot.len());
-        context.record(LocalCopyRecord::new(
-            record_path.clone(),
-            LocalCopyAction::SkippedExisting,
-            0,
-            total_bytes,
-            Duration::default(),
-            Some(metadata_snapshot),
-        ));
-        return Ok(());
-    }
-
-    let use_sparse_writes = context.sparse_enabled();
-    let partial_enabled = context.partial_enabled();
-    let inplace_enabled = context.inplace_enabled();
-    let checksum_enabled = context.checksum_enabled();
-    let size_only_enabled = context.size_only_enabled();
-    let append_allowed = context.append_enabled();
-    let append_verify = context.append_verify_enabled();
-    let whole_file_enabled = context.whole_file_enabled();
-    let compress_enabled = context.should_compress(record_path.as_path());
-    let relative_for_link = relative.unwrap_or(record_path.as_path());
-
-    if let Some(existing) = existing_metadata.as_ref() {
-        context.backup_existing_entry(destination, relative, existing.file_type())?;
-    }
-
-    if let Some(link_target) = context.link_dest_target(
-        relative_for_link,
-        source,
-        metadata,
-        &metadata_options,
-        size_only_enabled,
-        checksum_enabled,
-    )? {
-        let mut attempted_commit = false;
-        loop {
-            match fs::hard_link(&link_target, destination) {
-                Ok(()) => break,
-                Err(error) if error.kind() == io::ErrorKind::AlreadyExists => {
-                    remove_existing_destination(destination)?;
-                    fs::hard_link(&link_target, destination).map_err(|link_error| {
-                        LocalCopyError::io(
-                            "create hard link",
-                            destination.to_path_buf(),
-                            link_error,
-                        )
-                    })?;
-                    break;
-                }
-                Err(error)
-                    if error.kind() == io::ErrorKind::NotFound
-                        && context.delay_updates_enabled()
-                        && !attempted_commit =>
-                {
-                    context.commit_deferred_update_for(&link_target)?;
-                    attempted_commit = true;
-                    continue;
-                }
-                Err(error) => {
-                    return Err(LocalCopyError::io(
-                        "create hard link",
-                        destination.to_path_buf(),
-                        error,
-                    ));
-                }
-            }
-        }
-
-        context.record_hard_link(metadata, destination);
-        context.summary_mut().record_hard_link();
-        let metadata_snapshot = LocalCopyMetadata::from_metadata(metadata, None);
-        let total_bytes = Some(metadata_snapshot.len());
-        context.record(LocalCopyRecord::new(
-            record_path.clone(),
-            LocalCopyAction::HardLink,
-            0,
-            total_bytes,
-            Duration::default(),
-            Some(metadata_snapshot),
-        ));
-        context.register_created_path(
-            destination,
-            CreatedEntryKind::HardLink,
-            destination_previously_existed,
-        );
-        remove_source_entry_if_requested(context, source, Some(record_path.as_path()), file_type)?;
-        return Ok(());
-    }
-    let mut copy_source_override: Option<PathBuf> = None;
-
-    if let Some(existing_target) = context.existing_hard_link_target(metadata) {
-        let mut attempted_commit = false;
-        loop {
-            match create_hard_link(&existing_target, destination) {
-                Ok(()) => break,
-                Err(error) if error.kind() == io::ErrorKind::AlreadyExists => {
-                    remove_existing_destination(destination)?;
-                    create_hard_link(&existing_target, destination).map_err(|link_error| {
-                        LocalCopyError::io(
-                            "create hard link",
-                            destination.to_path_buf(),
-                            link_error,
-                        )
-                    })?;
-                    break;
-                }
-                Err(error)
-                    if error.kind() == io::ErrorKind::NotFound
-                        && context.delay_updates_enabled()
-                        && !attempted_commit =>
-                {
-                    context.commit_deferred_update_for(&existing_target)?;
-                    attempted_commit = true;
-                    continue;
-                }
-                Err(error) => {
-                    return Err(LocalCopyError::io(
-                        "create hard link",
-                        destination.to_path_buf(),
-                        error,
-                    ));
-                }
-            }
-        }
-
-        context.record_hard_link(metadata, destination);
-        context.summary_mut().record_hard_link();
-        let metadata_snapshot = LocalCopyMetadata::from_metadata(metadata, None);
-        let total_bytes = Some(metadata_snapshot.len());
-        context.record(LocalCopyRecord::new(
-            record_path.clone(),
-            LocalCopyAction::HardLink,
-            0,
-            total_bytes,
-            Duration::default(),
-            Some(metadata_snapshot),
-        ));
-        context.register_created_path(
-            destination,
-            CreatedEntryKind::HardLink,
-            destination_previously_existed,
-        );
-        remove_source_entry_if_requested(context, source, Some(record_path.as_path()), file_type)?;
-        return Ok(());
-    }
-
-    if !context.reference_directories().is_empty() && !record_path.as_os_str().is_empty() {
-        if let Some(decision) = find_reference_action(
-            context,
-            ReferenceQuery {
-                destination,
-                relative: record_path.as_path(),
-                source,
-                metadata,
-                metadata_options: &metadata_options,
-                size_only: size_only_enabled,
-                checksum: checksum_enabled,
-            },
-        )? {
-            match decision {
-                ReferenceDecision::Skip => {
-                    context.summary_mut().record_regular_file_matched();
-                    let metadata_snapshot = LocalCopyMetadata::from_metadata(metadata, None);
-                    let total_bytes = Some(metadata_snapshot.len());
-                    context.record(LocalCopyRecord::new(
-                        record_path.clone(),
-                        LocalCopyAction::MetadataReused,
-                        0,
-                        total_bytes,
-                        Duration::default(),
-                        Some(metadata_snapshot),
-                    ));
-                    context.register_progress();
-                    remove_source_entry_if_requested(
-                        context,
-                        source,
-                        Some(record_path.as_path()),
-                        file_type,
-                    )?;
-                    return Ok(());
-                }
-                ReferenceDecision::Copy(path) => {
-                    copy_source_override = Some(path);
-                }
-                ReferenceDecision::Link(path) => {
-                    if existing_metadata.is_some() {
-                        remove_existing_destination(destination)?;
-                    }
-
-                    let link_result = create_hard_link(&path, destination);
-                    let mut degrade_to_copy = false;
-                    match link_result {
-                        Ok(()) => {}
-                        Err(error) if error.kind() == io::ErrorKind::AlreadyExists => {
-                            remove_existing_destination(destination)?;
-                            create_hard_link(&path, destination).map_err(|link_error| {
-                                LocalCopyError::io(
-                                    "create hard link",
-                                    destination.to_path_buf(),
-                                    link_error,
-                                )
-                            })?;
-                        }
-                        Err(error)
-                            if matches!(
-                                error.raw_os_error(),
-                                Some(code) if code == CROSS_DEVICE_ERROR_CODE
-                            ) =>
-                        {
-                            degrade_to_copy = true;
-                        }
-                        Err(error) => {
-                            return Err(LocalCopyError::io(
-                                "create hard link",
-                                destination.to_path_buf(),
-                                error,
-                            ));
-                        }
-                    }
-
-                    if degrade_to_copy {
-                        copy_source_override = Some(path);
-                    } else if copy_source_override.is_none() {
-                        apply_file_metadata_with_options(
-                            destination,
-                            metadata,
-                            metadata_options.clone(),
-                        )
-                        .map_err(map_metadata_error)?;
-                        #[cfg(feature = "xattr")]
-                        sync_xattrs_if_requested(preserve_xattrs, mode, source, destination, true)?;
-                        #[cfg(feature = "acl")]
-                        sync_acls_if_requested(preserve_acls, mode, source, destination, true)?;
-                        context.record_hard_link(metadata, destination);
-                        context.summary_mut().record_hard_link();
-                        let metadata_snapshot = LocalCopyMetadata::from_metadata(metadata, None);
-                        let total_bytes = Some(metadata_snapshot.len());
-                        context.record(LocalCopyRecord::new(
-                            record_path.clone(),
-                            LocalCopyAction::HardLink,
-                            0,
-                            total_bytes,
-                            Duration::default(),
-                            Some(metadata_snapshot),
-                        ));
-                        context.register_created_path(
-                            destination,
-                            CreatedEntryKind::HardLink,
-                            destination_previously_existed,
-                        );
-                        context.register_progress();
-                        remove_source_entry_if_requested(
-                            context,
-                            source,
-                            Some(record_path.as_path()),
-                            file_type,
-                        )?;
-                        return Ok(());
-                    }
-                }
-            }
-        }
-    }
-
-    if let Some(existing) = existing_metadata.as_ref() {
-        if should_skip_copy(CopyComparison {
-            source_path: source,
-            source: metadata,
-            destination_path: destination,
-            destination: existing,
-            options: &metadata_options,
-            size_only: size_only_enabled,
-            checksum: checksum_enabled,
-            checksum_algorithm: context.options().checksum_algorithm(),
-            modify_window: context.options().modify_window(),
-        }) {
-            apply_file_metadata_with_options(destination, metadata, metadata_options.clone())
-                .map_err(map_metadata_error)?;
-            #[cfg(feature = "xattr")]
-            sync_xattrs_if_requested(preserve_xattrs, mode, source, destination, true)?;
-            #[cfg(feature = "acl")]
-            sync_acls_if_requested(preserve_acls, mode, source, destination, true)?;
-            context.record_hard_link(metadata, destination);
-            context.summary_mut().record_regular_file_matched();
-            let metadata_snapshot = LocalCopyMetadata::from_metadata(metadata, None);
-            let total_bytes = Some(metadata_snapshot.len());
-            context.record(LocalCopyRecord::new(
-                record_path.clone(),
-                LocalCopyAction::MetadataReused,
-                0,
-                total_bytes,
-                Duration::default(),
-                Some(metadata_snapshot),
-            ));
-            return Ok(());
-        }
-    }
-
-    let mut reader = fs::File::open(source)
-        .map_err(|error| LocalCopyError::io("copy file", source.to_path_buf(), error))?;
-    let append_mode = determine_append_mode(
-        append_allowed,
-        append_verify,
-        &mut reader,
-        source,
-        destination,
-        existing_metadata.as_ref(),
-        file_size,
-    )?;
-    let append_offset = match append_mode {
-        AppendMode::Append(offset) => offset,
-        AppendMode::Disabled => 0,
-    };
-    reader
-        .seek(SeekFrom::Start(append_offset))
-        .map_err(|error| LocalCopyError::io("copy file", source.to_path_buf(), error))?;
-    let delta_signature = if append_offset == 0 && !whole_file_enabled && !inplace_enabled {
-        match existing_metadata.as_ref() {
-            Some(existing) if existing.is_file() => build_delta_signature(destination, existing)?,
-            _ => None,
-        }
-    } else {
-        None
-    };
-
-    let copy_source = copy_source_override.as_deref().unwrap_or(source);
-    let mut reader = fs::File::open(copy_source)
-        .map_err(|error| LocalCopyError::io("copy file", copy_source.to_path_buf(), error))?;
-    if append_offset > 0 {
-        reader
-            .seek(SeekFrom::Start(append_offset))
-            .map_err(|error| LocalCopyError::io("copy file", copy_source.to_path_buf(), error))?;
-    }
-    let mut guard = None;
-    let mut staging_path: Option<PathBuf> = None;
-
-    let mut writer = if append_offset > 0 {
-        let mut file = fs::OpenOptions::new()
-            .create(true)
-            .write(true)
-            .truncate(false)
-            .open(destination)
-            .map_err(|error| LocalCopyError::io("copy file", destination.to_path_buf(), error))?;
-        file.seek(SeekFrom::Start(append_offset))
-            .map_err(|error| LocalCopyError::io("copy file", destination.to_path_buf(), error))?;
-        file
-    } else if inplace_enabled {
-        fs::OpenOptions::new()
-            .create(true)
-            .write(true)
-            .truncate(true)
-            .open(destination)
-            .map_err(|error| LocalCopyError::io("copy file", destination.to_path_buf(), error))?
-    } else {
-        let (new_guard, file) = DestinationWriteGuard::new(
-            destination,
-            partial_enabled,
-            context.partial_directory_path(),
-            context.temp_directory_path(),
-        )?;
-        staging_path = Some(new_guard.staging_path().to_path_buf());
-        guard = Some(new_guard);
-        file
-    };
-    let preallocate_target = guard
-        .as_ref()
-        .map(|existing_guard| existing_guard.staging_path())
-        .unwrap_or(destination);
-    maybe_preallocate_destination(
-        &mut writer,
-        preallocate_target,
-        file_size,
-        append_offset,
-        context.preallocate_enabled(),
-    )?;
-    let mut buffer = vec![0u8; COPY_BUFFER_SIZE];
-
-    let start = Instant::now();
-
-    let copy_result = context.copy_file_contents(
-        &mut reader,
-        &mut writer,
-        &mut buffer,
-        use_sparse_writes,
-        compress_enabled,
-        source,
-        destination,
-        record_path.as_path(),
-        delta_signature.as_ref(),
-        file_size,
-        append_offset,
-        start,
-    );
-
-    drop(writer);
-
-    let staging_path_for_links = guard
-        .as_ref()
-        .map(|existing_guard| existing_guard.staging_path().to_path_buf())
-        .or_else(|| staging_path.take());
-    let delay_updates_enabled = context.delay_updates_enabled();
-
-    let outcome = match copy_result {
-        Ok(outcome) => {
-            if let Err(timeout_error) = context.enforce_timeout() {
-                if let Some(guard) = guard.take() {
-                    guard.discard();
-                }
-
-                if existing_metadata.is_none() {
-                    remove_incomplete_destination(destination);
-                }
-
-                return Err(timeout_error);
-            }
-            outcome
-        }
-        Err(error) => {
-            if let Some(guard) = guard.take() {
-                guard.discard();
-            }
-
-            if existing_metadata.is_none() {
-                remove_incomplete_destination(destination);
-            }
-
-            return Err(error);
-        }
-    };
-
-    context.register_created_path(
-        destination,
-        CreatedEntryKind::File,
-        destination_previously_existed,
-    );
-
-    let hard_link_path = if delay_updates_enabled {
-        staging_path_for_links.as_deref().unwrap_or(destination)
-    } else {
-        destination
-    };
-    context.record_hard_link(metadata, hard_link_path);
-    let elapsed = start.elapsed();
-    let compressed_bytes = outcome.compressed_bytes();
-    context
-        .summary_mut()
-        .record_file(file_size, outcome.literal_bytes(), compressed_bytes);
-    context.summary_mut().record_elapsed(elapsed);
-    let metadata_snapshot = LocalCopyMetadata::from_metadata(metadata, None);
-    let total_bytes = Some(metadata_snapshot.len());
-    context.record(
-        LocalCopyRecord::new(
-            record_path.clone(),
-            LocalCopyAction::DataCopied,
-            outcome.literal_bytes(),
-            total_bytes,
-            elapsed,
-            Some(metadata_snapshot),
-        )
-        .with_creation(!destination_previously_existed),
-    );
-
-    if let Err(timeout_error) = context.enforce_timeout() {
-        if existing_metadata.is_none() {
-            remove_incomplete_destination(destination);
-        }
-
-        return Err(timeout_error);
-    }
-
-    let relative_for_removal = Some(record_path.clone());
-    if let Some(guard) = guard {
-        if delay_updates_enabled {
-            let destination_path = guard.final_path().to_path_buf();
-            let update = DeferredUpdate::new(
-                guard,
-                metadata.clone(),
-                metadata_options.clone(),
-                mode,
-                source.to_path_buf(),
-                relative_for_removal.clone(),
-                destination_path,
-                file_type,
-                destination_previously_existed,
-                #[cfg(feature = "xattr")]
-                preserve_xattrs,
-                #[cfg(feature = "acl")]
-                preserve_acls,
-            );
-            context.register_deferred_update(update);
-        } else {
-            let destination_path = guard.final_path().to_path_buf();
-            guard.commit()?;
-            context.apply_metadata_and_finalize(
-                destination_path.as_path(),
-                FinalizeMetadataParams::new(
-                    metadata,
-                    metadata_options.clone(),
-                    mode,
-                    source,
-                    relative_for_removal.as_deref(),
-                    file_type,
-                    destination_previously_existed,
-                    #[cfg(feature = "xattr")]
-                    preserve_xattrs,
-                    #[cfg(feature = "acl")]
-                    preserve_acls,
-                ),
-            )?;
-        }
-    } else {
-        context.apply_metadata_and_finalize(
-            destination,
-            FinalizeMetadataParams::new(
-                metadata,
-                metadata_options,
-                mode,
-                source,
-                relative_for_removal.as_deref(),
-                file_type,
-                destination_previously_existed,
-                #[cfg(feature = "xattr")]
-                preserve_xattrs,
-                #[cfg(feature = "acl")]
-                preserve_acls,
-            ),
-        )?;
-    }
-
-    Ok(())
-}
-
-enum AppendMode {
-    Disabled,
-    Append(u64),
-}
-
-fn determine_append_mode(
-    append_allowed: bool,
-    append_verify: bool,
-    reader: &mut fs::File,
-    source: &Path,
-    destination: &Path,
-    existing_metadata: Option<&fs::Metadata>,
-    file_size: u64,
-) -> Result<AppendMode, LocalCopyError> {
-    if !append_allowed {
-        return Ok(AppendMode::Disabled);
-    }
-
-    let existing = match existing_metadata {
-        Some(meta) if meta.is_file() => meta,
-        _ => return Ok(AppendMode::Disabled),
-    };
-
-    let existing_len = existing.len();
-    if existing_len == 0 || existing_len >= file_size {
-        reader
-            .seek(SeekFrom::Start(0))
-            .map_err(|error| LocalCopyError::io("copy file", source.to_path_buf(), error))?;
-        return Ok(AppendMode::Disabled);
-    }
-
-    if append_verify {
-        let matches = verify_append_prefix(reader, source, destination, existing_len)?;
-        reader
-            .seek(SeekFrom::Start(0))
-            .map_err(|error| LocalCopyError::io("copy file", source.to_path_buf(), error))?;
-        if !matches {
-            return Ok(AppendMode::Disabled);
-        }
-    } else {
-        reader
-            .seek(SeekFrom::Start(0))
-            .map_err(|error| LocalCopyError::io("copy file", source.to_path_buf(), error))?;
-    }
-
-    Ok(AppendMode::Append(existing_len))
-}
-
-fn verify_append_prefix(
-    reader: &mut fs::File,
-    source: &Path,
-    destination: &Path,
-    existing_len: u64,
-) -> Result<bool, LocalCopyError> {
-    reader
-        .seek(SeekFrom::Start(0))
-        .map_err(|error| LocalCopyError::io("copy file", source.to_path_buf(), error))?;
-    let mut destination_file = fs::File::open(destination).map_err(|error| {
-        LocalCopyError::io(
-            "read existing destination",
-            destination.to_path_buf(),
-            error,
-        )
-    })?;
-    let mut remaining = existing_len;
-    let mut source_buffer = vec![0u8; COPY_BUFFER_SIZE];
-    let mut destination_buffer = vec![0u8; COPY_BUFFER_SIZE];
-
-    while remaining > 0 {
-        let chunk = remaining.min(COPY_BUFFER_SIZE as u64) as usize;
-        let source_read = reader
-            .read(&mut source_buffer[..chunk])
-            .map_err(|error| LocalCopyError::io("copy file", source.to_path_buf(), error))?;
-        let destination_read = destination_file
-            .read(&mut destination_buffer[..chunk])
-            .map_err(|error| {
-                LocalCopyError::io(
-                    "read existing destination",
-                    destination.to_path_buf(),
-                    error,
-                )
-            })?;
-
-        if source_read == 0 || destination_read == 0 || source_read != destination_read {
-            return Ok(false);
-        }
-
-        if source_buffer[..source_read] != destination_buffer[..destination_read] {
-            return Ok(false);
-        }
-
-        remaining = remaining.saturating_sub(source_read as u64);
-    }
-
-    Ok(true)
-}
-
-fn partial_destination_path(destination: &Path) -> PathBuf {
-    let file_name = destination
-        .file_name()
-        .map(|name| name.to_string_lossy().to_string())
-        .unwrap_or_else(|| "partial".to_string());
-    let partial_name = format!(".rsync-partial-{}", file_name);
-    destination.with_file_name(partial_name)
-}
-
-fn partial_directory_destination_path(
-    destination: &Path,
-    partial_dir: &Path,
-) -> Result<PathBuf, LocalCopyError> {
-    let base_dir = if partial_dir.is_absolute() {
-        partial_dir.to_path_buf()
-    } else {
-        let parent = destination
-            .parent()
-            .map_or_else(|| PathBuf::from("."), Path::to_path_buf);
-        parent.join(partial_dir)
-    };
-    fs::create_dir_all(&base_dir)
-        .map_err(|error| LocalCopyError::io("create partial directory", base_dir.clone(), error))?;
-    let file_name = destination
-        .file_name()
-        .map(|name| name.to_os_string())
-        .unwrap_or_else(|| OsStr::new("partial").to_os_string());
-    Ok(base_dir.join(file_name))
-}
-
-fn compute_backup_path(
-    destination_root: &Path,
-    destination: &Path,
-    relative: Option<&Path>,
-    backup_dir: Option<&Path>,
-    suffix: &OsStr,
-) -> PathBuf {
-    let relative_path = if let Some(rel) = relative {
-        rel.to_path_buf()
-    } else if let Ok(stripped) = destination.strip_prefix(destination_root) {
-        if stripped.as_os_str().is_empty() {
-            destination
-                .file_name()
-                .map(PathBuf::from)
-                .unwrap_or_else(|| PathBuf::from(destination))
-        } else {
-            stripped.to_path_buf()
-        }
-    } else if let Some(name) = destination.file_name() {
-        PathBuf::from(name)
-    } else {
-        PathBuf::from(destination)
-    };
-
-    let mut backup_name = relative_path
-        .file_name()
-        .map(|name| name.to_os_string())
-        .unwrap_or_else(|| OsString::from("backup"));
-    if !suffix.is_empty() {
-        backup_name.push(suffix);
-    }
-
-    let mut base = if let Some(dir) = backup_dir {
-        let mut base = if dir.is_absolute() {
-            dir.to_path_buf()
-        } else {
-            destination_root.join(dir)
-        };
-        if let Some(parent) = relative_path.parent() {
-            if !parent.as_os_str().is_empty() {
-                base = base.join(parent);
-            }
-        }
-        base
-    } else {
-        destination
-            .parent()
-            .map(Path::to_path_buf)
-            .unwrap_or_else(|| PathBuf::from("."))
-    };
-
-    base.push(backup_name);
-    base
-}
-
-fn copy_entry_to_backup(
-    source: &Path,
-    backup_path: &Path,
-    file_type: fs::FileType,
-) -> Result<(), LocalCopyError> {
-    if file_type.is_file() {
-        fs::copy(source, backup_path).map_err(|error| {
-            LocalCopyError::io("create backup", backup_path.to_path_buf(), error)
-        })?;
-    } else if file_type.is_symlink() {
-        let target = fs::read_link(source).map_err(|error| {
-            LocalCopyError::io("read symbolic link", source.to_path_buf(), error)
-        })?;
-        create_symlink(&target, source, backup_path).map_err(|error| {
-            LocalCopyError::io("create symbolic link", backup_path.to_path_buf(), error)
-        })?;
-    }
-    Ok(())
-}
-
-fn remove_existing_destination(path: &Path) -> Result<(), LocalCopyError> {
-    match fs::remove_file(path) {
-        Ok(()) => Ok(()),
-        Err(error) if error.kind() == io::ErrorKind::NotFound => Ok(()),
-        Err(error) => Err(LocalCopyError::io(
-            "remove existing destination",
-            path.to_path_buf(),
-            error,
-        )),
-    }
-}
-
-fn temporary_destination_path(
-    destination: &Path,
-    unique: usize,
-    temp_dir: Option<&Path>,
-) -> PathBuf {
-    let file_name = destination
-        .file_name()
-        .map(|name| name.to_string_lossy().to_string())
-        .unwrap_or_else(|| "temp".to_string());
-    let temp_name = format!(".rsync-tmp-{file_name}-{}-{}", process::id(), unique);
-    match temp_dir {
-        Some(dir) => dir.join(temp_name),
-        None => destination.with_file_name(temp_name),
-    }
-}
-
-pub(crate) struct DestinationWriteGuard {
-    final_path: PathBuf,
-    temp_path: PathBuf,
-    preserve_on_error: bool,
-    committed: bool,
-}
-
-impl DestinationWriteGuard {
-    fn new(
-        destination: &Path,
-        partial: bool,
-        partial_dir: Option<&Path>,
-        temp_dir: Option<&Path>,
-    ) -> Result<(Self, fs::File), LocalCopyError> {
-        if partial {
-            let temp_path = if let Some(dir) = partial_dir {
-                partial_directory_destination_path(destination, dir)?
-            } else {
-                partial_destination_path(destination)
-            };
-            if let Err(error) = fs::remove_file(&temp_path) {
-                if error.kind() != io::ErrorKind::NotFound {
-                    return Err(LocalCopyError::io(
-                        "remove existing partial file",
-                        temp_path.clone(),
-                        error,
-                    ));
-                }
-            }
-            let file = fs::OpenOptions::new()
-                .create(true)
-                .write(true)
-                .truncate(true)
-                .open(&temp_path)
-                .map_err(|error| LocalCopyError::io("copy file", temp_path.clone(), error))?;
-            Ok((
-                Self {
-                    final_path: destination.to_path_buf(),
-                    temp_path,
-                    preserve_on_error: true,
-                    committed: false,
-                },
-                file,
-            ))
-        } else {
-            loop {
-                let unique = NEXT_TEMP_FILE_ID.fetch_add(1, AtomicOrdering::Relaxed);
-                let temp_path = temporary_destination_path(destination, unique, temp_dir);
-                match fs::OpenOptions::new()
-                    .write(true)
-                    .create_new(true)
-                    .open(&temp_path)
-                {
-                    Ok(file) => {
-                        return Ok((
-                            Self {
-                                final_path: destination.to_path_buf(),
-                                temp_path,
-                                preserve_on_error: false,
-                                committed: false,
-                            },
-                            file,
-                        ));
-                    }
-                    Err(error) if error.kind() == io::ErrorKind::AlreadyExists => {
-                        continue;
-                    }
-                    Err(error) => {
-                        return Err(LocalCopyError::io("copy file", temp_path.clone(), error));
-                    }
-                }
-            }
-        }
-    }
-
-    fn staging_path(&self) -> &Path {
-        &self.temp_path
-    }
-
-    fn commit(mut self) -> Result<(), LocalCopyError> {
-        match fs::rename(&self.temp_path, &self.final_path) {
-            Ok(()) => {}
-            Err(error) if error.kind() == io::ErrorKind::AlreadyExists => {
-                remove_existing_destination(&self.final_path)?;
-                fs::rename(&self.temp_path, &self.final_path).map_err(|rename_error| {
-                    LocalCopyError::io(self.finalise_action(), self.temp_path.clone(), rename_error)
-                })?;
-            }
-            Err(error) if error.kind() == io::ErrorKind::CrossesDevices => {
-                fs::copy(&self.temp_path, &self.final_path).map_err(|copy_error| {
-                    LocalCopyError::io(self.finalise_action(), self.final_path.clone(), copy_error)
-                })?;
-                fs::remove_file(&self.temp_path).map_err(|remove_error| {
-                    LocalCopyError::io(self.finalise_action(), self.temp_path.clone(), remove_error)
-                })?;
-            }
-            Err(error) => {
-                return Err(LocalCopyError::io(
-                    self.finalise_action(),
-                    self.temp_path.clone(),
-                    error,
-                ));
-            }
-        }
-        self.committed = true;
-        Ok(())
-    }
-
-    fn final_path(&self) -> &Path {
-        &self.final_path
-    }
-
-    fn discard(mut self) {
-        if self.preserve_on_error {
-            self.committed = true;
-            return;
-        }
-
-        if let Err(error) = fs::remove_file(&self.temp_path) {
-            if error.kind() != io::ErrorKind::NotFound {
-                // Best-effort cleanup: the file may have been removed concurrently.
-            }
-        }
-
-        self.committed = true;
-    }
-
-    fn finalise_action(&self) -> &'static str {
-        if self.preserve_on_error {
-            "finalise partial file"
-        } else {
-            "finalise temporary file"
-        }
-    }
-}
-
-impl Drop for DestinationWriteGuard {
-    fn drop(&mut self) {
-        if !self.committed && !self.preserve_on_error {
-            let _ = fs::remove_file(&self.temp_path);
-        }
-    }
-}
-
-fn remove_incomplete_destination(destination: &Path) {
-    if let Err(error) = fs::remove_file(destination) {
-        if error.kind() != io::ErrorKind::NotFound {
-            // Preserve the original error from the transfer attempt.
-        }
-    }
-}
-
-fn write_sparse_chunk(
-    writer: &mut fs::File,
-    chunk: &[u8],
-    destination: &Path,
-) -> Result<usize, LocalCopyError> {
-    let mut index = 0usize;
-    let mut written = 0usize;
-
-    while index < chunk.len() {
-        if chunk[index] == 0 {
-            let start = index;
-            while index < chunk.len() && chunk[index] == 0 {
-                index += 1;
-            }
-            let span = index - start;
-            if span > 0 {
-                writer
-                    .seek(SeekFrom::Current(span as i64))
-                    .map_err(|error| {
-                        LocalCopyError::io(
-                            "seek in destination file",
-                            destination.to_path_buf(),
-                            error,
-                        )
-                    })?;
-            }
-        } else {
-            let start = index;
-            while index < chunk.len() && chunk[index] != 0 {
-                index += 1;
-            }
-            writer.write_all(&chunk[start..index]).map_err(|error| {
-                LocalCopyError::io("copy file", destination.to_path_buf(), error)
-            })?;
-            written = written.saturating_add(index - start);
-        }
-    }
-
-    Ok(written)
-}
-
-fn destination_is_newer(source: &fs::Metadata, destination: &fs::Metadata) -> bool {
-    match (source.modified(), destination.modified()) {
-        (Ok(src), Ok(dst)) => dst > src,
-        _ => false,
-    }
-}
-
-fn build_delta_signature(
-    destination: &Path,
-    metadata: &fs::Metadata,
-) -> Result<Option<DeltaSignatureIndex>, LocalCopyError> {
-    let length = metadata.len();
-    if length == 0 {
-        return Ok(None);
-    }
-
-    let checksum_len = NonZeroU8::new(16).expect("strong checksum length must be non-zero");
-    let params = SignatureLayoutParams::new(length, None, ProtocolVersion::NEWEST, checksum_len);
-    let layout = match calculate_signature_layout(params) {
-        Ok(layout) => layout,
-        Err(_) => return Ok(None),
-    };
-
-    let signature = match generate_file_signature(
-        fs::File::open(destination).map_err(|error| {
-            LocalCopyError::io(
-                "read existing destination",
-                destination.to_path_buf(),
-                error,
-            )
-        })?,
-        layout,
-        SignatureAlgorithm::Md4,
-    ) {
-        Ok(signature) => signature,
-        Err(SignatureError::Io(error)) => {
-            return Err(LocalCopyError::io(
-                "read existing destination",
-                destination.to_path_buf(),
-                error,
-            ));
-        }
-        Err(_) => return Ok(None),
-    };
-
-    match DeltaSignatureIndex::from_signature(&signature, SignatureAlgorithm::Md4) {
-        Some(index) => Ok(Some(index)),
-        None => Ok(None),
-    }
-}
-
-struct CopyComparison<'a> {
-    source_path: &'a Path,
-    source: &'a fs::Metadata,
-    destination_path: &'a Path,
-    destination: &'a fs::Metadata,
-    options: &'a MetadataOptions,
-    size_only: bool,
-    checksum: bool,
-    checksum_algorithm: SignatureAlgorithm,
-    modify_window: Duration,
-}
-
-fn should_skip_copy(params: CopyComparison<'_>) -> bool {
-    let CopyComparison {
-        source_path,
-        source,
-        destination_path,
-        destination,
-        options,
-        size_only,
-        checksum,
-        checksum_algorithm,
-        modify_window,
-    } = params;
-    if destination.len() != source.len() {
-        return false;
-    }
-
-    if checksum {
-        return files_checksum_match(source_path, destination_path, checksum_algorithm)
-            .unwrap_or(false);
-    }
-
-    if size_only {
-        return true;
-    }
-
-    if options.times() {
-        match (source.modified(), destination.modified()) {
-            (Ok(src), Ok(dst)) if system_time_within_window(src, dst, modify_window) => {}
-            _ => return false,
-        }
-    } else {
-        return false;
-    }
-
-    files_match(source_path, destination_path)
-}
-
-fn system_time_within_window(a: SystemTime, b: SystemTime, window: Duration) -> bool {
-    if window.is_zero() {
-        return a.eq(&b);
-    }
-
-    match a.duration_since(b) {
-        Ok(diff) => diff <= window,
-        Err(_) => matches!(b.duration_since(a), Ok(diff) if diff <= window),
-    }
-}
-
-fn files_match(source: &Path, destination: &Path) -> bool {
-    let mut source_file = match fs::File::open(source) {
-        Ok(file) => file,
-        Err(_) => return false,
-    };
-    let mut destination_file = match fs::File::open(destination) {
-        Ok(file) => file,
-        Err(_) => return false,
-    };
-
-    let mut source_buffer = vec![0u8; COPY_BUFFER_SIZE];
-    let mut destination_buffer = vec![0u8; COPY_BUFFER_SIZE];
-
-    loop {
-        let source_read = match source_file.read(&mut source_buffer) {
-            Ok(bytes) => bytes,
-            Err(_) => return false,
-        };
-        let destination_read = match destination_file.read(&mut destination_buffer) {
-            Ok(bytes) => bytes,
-            Err(_) => return false,
-        };
-
-        if source_read != destination_read {
-            return false;
-        }
-
-        if source_read == 0 {
-            return true;
-        }
-
-        if source_buffer[..source_read] != destination_buffer[..destination_read] {
-            return false;
-        }
-    }
-}
-
-enum StrongHasher {
-    Md4(Md4),
-    Md5(Md5),
-    Xxh64(Xxh64),
-    Xxh3(Xxh3),
-    Xxh128(Xxh3_128),
-}
-
-impl StrongHasher {
-    fn new(algorithm: SignatureAlgorithm) -> Self {
-        match algorithm {
-            SignatureAlgorithm::Md4 => StrongHasher::Md4(Md4::new()),
-            SignatureAlgorithm::Md5 => StrongHasher::Md5(Md5::new()),
-            SignatureAlgorithm::Xxh64 { seed } => StrongHasher::Xxh64(Xxh64::new(seed)),
-            SignatureAlgorithm::Xxh3 { seed } => StrongHasher::Xxh3(Xxh3::new(seed)),
-            SignatureAlgorithm::Xxh3_128 { seed } => StrongHasher::Xxh128(Xxh3_128::new(seed)),
-        }
-    }
-
-    fn update(&mut self, data: &[u8]) {
-        match self {
-            StrongHasher::Md4(state) => state.update(data),
-            StrongHasher::Md5(state) => state.update(data),
-            StrongHasher::Xxh64(state) => state.update(data),
-            StrongHasher::Xxh3(state) => state.update(data),
-            StrongHasher::Xxh128(state) => state.update(data),
-        }
-    }
-
-    fn finalize(self) -> Vec<u8> {
-        match self {
-            StrongHasher::Md4(state) => state.finalize().as_ref().to_vec(),
-            StrongHasher::Md5(state) => state.finalize().as_ref().to_vec(),
-            StrongHasher::Xxh64(state) => state.finalize().as_ref().to_vec(),
-            StrongHasher::Xxh3(state) => state.finalize().as_ref().to_vec(),
-            StrongHasher::Xxh128(state) => state.finalize().as_ref().to_vec(),
-        }
-    }
-}
-
-fn files_checksum_match(
-    source: &Path,
-    destination: &Path,
-    algorithm: SignatureAlgorithm,
-) -> io::Result<bool> {
-    let mut source_file = fs::File::open(source)?;
-    let mut destination_file = fs::File::open(destination)?;
-
-    let mut source_hasher = StrongHasher::new(algorithm);
-    let mut destination_hasher = StrongHasher::new(algorithm);
-
-    let mut source_buffer = vec![0u8; COPY_BUFFER_SIZE];
-    let mut destination_buffer = vec![0u8; COPY_BUFFER_SIZE];
-
-    loop {
-        let source_read = source_file.read(&mut source_buffer)?;
-        let destination_read = destination_file.read(&mut destination_buffer)?;
-
-        if source_read != destination_read {
-            return Ok(false);
-        }
-
-        if source_read == 0 {
-            break;
-        }
-
-        source_hasher.update(&source_buffer[..source_read]);
-        destination_hasher.update(&destination_buffer[..destination_read]);
-    }
-
-    Ok(source_hasher.finalize() == destination_hasher.finalize())
-}
-
-fn copy_fifo(
-    context: &mut CopyContext,
-    source: &Path,
-    destination: &Path,
-    metadata: &fs::Metadata,
-    metadata_options: &MetadataOptions,
-    relative: Option<&Path>,
-) -> Result<(), LocalCopyError> {
-    context.enforce_timeout()?;
-    let mode = context.mode();
-    let file_type = metadata.file_type();
-    #[cfg(feature = "xattr")]
-    let preserve_xattrs = context.xattrs_enabled();
-    #[cfg(feature = "acl")]
-    let preserve_acls = context.acls_enabled();
-    let record_path = relative
-        .map(Path::to_path_buf)
-        .or_else(|| destination.file_name().map(PathBuf::from));
-    context.summary_mut().record_fifo_total();
-    if let Some(parent) = destination.parent() {
-        if !parent.as_os_str().is_empty() {
-            if mode.is_dry_run() {
-                match fs::symlink_metadata(parent) {
-                    Ok(existing) if !existing.file_type().is_dir() => {
-                        return Err(LocalCopyError::invalid_argument(
-                            LocalCopyArgumentError::ReplaceNonDirectoryWithDirectory,
-                        ));
-                    }
-                    Ok(_) => {}
-                    Err(error) if error.kind() == io::ErrorKind::NotFound => {}
-                    Err(error) => {
-                        return Err(LocalCopyError::io(
-                            "inspect existing destination",
-                            parent.to_path_buf(),
-                            error,
-                        ));
-                    }
-                }
-            } else {
-                fs::create_dir_all(parent).map_err(|error| {
-                    LocalCopyError::io("create parent directory", parent.to_path_buf(), error)
-                })?;
-                context.register_progress();
-            }
-        }
-    }
-
-    if mode.is_dry_run() {
-        match fs::symlink_metadata(destination) {
-            Ok(existing) => {
-                if existing.file_type().is_dir() {
-                    return Err(LocalCopyError::invalid_argument(
-                        LocalCopyArgumentError::ReplaceDirectoryWithSpecial,
-                    ));
-                }
-            }
-            Err(error) if error.kind() == io::ErrorKind::NotFound => {}
-            Err(error) => {
-                return Err(LocalCopyError::io(
-                    "inspect existing destination",
-                    destination.to_path_buf(),
-                    error,
-                ));
-            }
-        }
-
-        context.summary_mut().record_fifo();
-        if let Some(path) = &record_path {
-            let metadata_snapshot = LocalCopyMetadata::from_metadata(metadata, None);
-            let total_bytes = Some(metadata_snapshot.len());
-            context.record(LocalCopyRecord::new(
-                path.clone(),
-                LocalCopyAction::FifoCopied,
-                0,
-                total_bytes,
-                Duration::default(),
-                Some(metadata_snapshot),
-            ));
-        }
-        context.register_progress();
-        remove_source_entry_if_requested(context, source, record_path.as_deref(), file_type)?;
-        return Ok(());
-    }
-
-    let mut destination_previously_existed = false;
-    match fs::symlink_metadata(destination) {
-        Ok(existing) => {
-            destination_previously_existed = true;
-            if existing.file_type().is_dir() {
-                return Err(LocalCopyError::invalid_argument(
-                    LocalCopyArgumentError::ReplaceDirectoryWithSpecial,
-                ));
-            }
-
-            context.backup_existing_entry(destination, relative, existing.file_type())?;
-            remove_existing_destination(destination)?;
-        }
-        Err(error) if error.kind() == io::ErrorKind::NotFound => {}
-        Err(error) => {
-            return Err(LocalCopyError::io(
-                "inspect existing destination",
-                destination.to_path_buf(),
-                error,
-            ));
-        }
-    }
-
-    create_fifo(destination, metadata).map_err(map_metadata_error)?;
-    context.register_created_path(
-        destination,
-        CreatedEntryKind::Fifo,
-        destination_previously_existed,
-    );
-    apply_file_metadata_with_options(destination, metadata, metadata_options.clone())
-        .map_err(map_metadata_error)?;
-    #[cfg(feature = "xattr")]
-    sync_xattrs_if_requested(preserve_xattrs, mode, source, destination, true)?;
-    #[cfg(feature = "acl")]
-    sync_acls_if_requested(preserve_acls, mode, source, destination, true)?;
-    context.summary_mut().record_fifo();
-    if let Some(path) = &record_path {
-        let metadata_snapshot = LocalCopyMetadata::from_metadata(metadata, None);
-        let total_bytes = Some(metadata_snapshot.len());
-        context.record(LocalCopyRecord::new(
-            path.clone(),
-            LocalCopyAction::FifoCopied,
-            0,
-            total_bytes,
-            Duration::default(),
-            Some(metadata_snapshot),
-        ));
-    }
-    context.register_progress();
-    remove_source_entry_if_requested(context, source, record_path.as_deref(), file_type)?;
-    Ok(())
-}
-
-fn copy_device(
-    context: &mut CopyContext,
-    source: &Path,
-    destination: &Path,
-    metadata: &fs::Metadata,
-    metadata_options: &MetadataOptions,
-    relative: Option<&Path>,
-) -> Result<(), LocalCopyError> {
-    context.enforce_timeout()?;
-    let mode = context.mode();
-    let file_type = metadata.file_type();
-    #[cfg(feature = "xattr")]
-    let preserve_xattrs = context.xattrs_enabled();
-    #[cfg(feature = "acl")]
-    let preserve_acls = context.acls_enabled();
-    let record_path = relative
-        .map(Path::to_path_buf)
-        .or_else(|| destination.file_name().map(PathBuf::from));
-    context.summary_mut().record_device_total();
-    if let Some(parent) = destination.parent() {
-        if !parent.as_os_str().is_empty() {
-            if mode.is_dry_run() {
-                match fs::symlink_metadata(parent) {
-                    Ok(existing) if !existing.file_type().is_dir() => {
-                        return Err(LocalCopyError::invalid_argument(
-                            LocalCopyArgumentError::ReplaceNonDirectoryWithDirectory,
-                        ));
-                    }
-                    Ok(_) => {}
-                    Err(error) if error.kind() == io::ErrorKind::NotFound => {}
-                    Err(error) => {
-                        return Err(LocalCopyError::io(
-                            "inspect existing destination",
-                            parent.to_path_buf(),
-                            error,
-                        ));
-                    }
-                }
-            } else {
-                fs::create_dir_all(parent).map_err(|error| {
-                    LocalCopyError::io("create parent directory", parent.to_path_buf(), error)
-                })?;
-                context.register_progress();
-            }
-        }
-    }
-
-    if mode.is_dry_run() {
-        match fs::symlink_metadata(destination) {
-            Ok(existing) => {
-                if existing.file_type().is_dir() {
-                    return Err(LocalCopyError::invalid_argument(
-                        LocalCopyArgumentError::ReplaceDirectoryWithSpecial,
-                    ));
-                }
-            }
-            Err(error) if error.kind() == io::ErrorKind::NotFound => {}
-            Err(error) => {
-                return Err(LocalCopyError::io(
-                    "inspect existing destination",
-                    destination.to_path_buf(),
-                    error,
-                ));
-            }
-        }
-
-        context.summary_mut().record_device();
-        if let Some(path) = &record_path {
-            let metadata_snapshot = LocalCopyMetadata::from_metadata(metadata, None);
-            let total_bytes = Some(metadata_snapshot.len());
-            context.record(LocalCopyRecord::new(
-                path.clone(),
-                LocalCopyAction::DeviceCopied,
-                0,
-                total_bytes,
-                Duration::default(),
-                Some(metadata_snapshot),
-            ));
-        }
-        context.register_progress();
-        remove_source_entry_if_requested(context, source, record_path.as_deref(), file_type)?;
-        return Ok(());
-    }
-
-    let mut destination_previously_existed = false;
-    match fs::symlink_metadata(destination) {
-        Ok(existing) => {
-            destination_previously_existed = true;
-            if existing.file_type().is_dir() {
-                return Err(LocalCopyError::invalid_argument(
-                    LocalCopyArgumentError::ReplaceDirectoryWithSpecial,
-                ));
-            }
-
-            context.backup_existing_entry(destination, relative, existing.file_type())?;
-            remove_existing_destination(destination)?;
-        }
-        Err(error) if error.kind() == io::ErrorKind::NotFound => {}
-        Err(error) => {
-            return Err(LocalCopyError::io(
-                "inspect existing destination",
-                destination.to_path_buf(),
-                error,
-            ));
-        }
-    }
-
-    create_device_node(destination, metadata).map_err(map_metadata_error)?;
-    context.register_created_path(
-        destination,
-        CreatedEntryKind::Device,
-        destination_previously_existed,
-    );
-    apply_file_metadata_with_options(destination, metadata, metadata_options.clone())
-        .map_err(map_metadata_error)?;
-    #[cfg(feature = "xattr")]
-    sync_xattrs_if_requested(preserve_xattrs, mode, source, destination, true)?;
-    #[cfg(feature = "acl")]
-    sync_acls_if_requested(preserve_acls, mode, source, destination, true)?;
-    context.summary_mut().record_device();
-    if let Some(path) = &record_path {
-        let metadata_snapshot = LocalCopyMetadata::from_metadata(metadata, None);
-        let total_bytes = Some(metadata_snapshot.len());
-        context.record(LocalCopyRecord::new(
-            path.clone(),
-            LocalCopyAction::DeviceCopied,
-            0,
-            total_bytes,
-            Duration::default(),
-            Some(metadata_snapshot),
-        ));
-    }
-    context.register_progress();
-    remove_source_entry_if_requested(context, source, record_path.as_deref(), file_type)?;
-    Ok(())
-}
-
-fn delete_extraneous_entries(
-    context: &mut CopyContext,
-    destination: &Path,
-    relative: Option<&Path>,
-    source_entries: &[OsString],
-) -> Result<(), LocalCopyError> {
-    let mut skipped_due_to_limit = 0u64;
-    let mut keep = HashSet::with_capacity(source_entries.len());
-    for name in source_entries {
-        keep.insert(name.clone());
-    }
-
-    let read_dir = match fs::read_dir(destination) {
-        Ok(iter) => iter,
-        Err(error) if error.kind() == io::ErrorKind::NotFound => return Ok(()),
-        Err(error) => {
-            return Err(LocalCopyError::io(
-                "read destination directory",
-                destination.to_path_buf(),
-                error,
-            ));
-        }
-    };
-
-    for entry in read_dir {
-        context.enforce_timeout()?;
-        let entry = entry.map_err(|error| {
-            LocalCopyError::io("read destination entry", destination.to_path_buf(), error)
-        })?;
-        let name = entry.file_name();
-
-        if keep.contains(&name) {
-            continue;
-        }
-
-        let name_path = PathBuf::from(name.as_os_str());
-        let path = destination.join(&name_path);
-        let entry_relative = match relative {
-            Some(base) => base.join(&name_path),
-            None => name_path.clone(),
-        };
-
-        let file_type = entry.file_type().map_err(|error| {
-            LocalCopyError::io("inspect extraneous destination entry", path.clone(), error)
-        })?;
-
-        if !context.allows_deletion(entry_relative.as_path(), file_type.is_dir()) {
-            continue;
-        }
-
-        if let Some(limit) = context.options().max_deletion_limit() {
-            if context.summary().items_deleted() >= limit {
-                skipped_due_to_limit = skipped_due_to_limit.saturating_add(1);
-                continue;
-            }
-        }
-
-        if context.mode().is_dry_run() {
-            context.summary_mut().record_deletion();
-            context.record(LocalCopyRecord::new(
-                entry_relative,
-                LocalCopyAction::EntryDeleted,
-                0,
-                None,
-                Duration::default(),
-                None,
-            ));
-            context.register_progress();
-            continue;
-        }
-
-        context.backup_existing_entry(&path, Some(entry_relative.as_path()), file_type)?;
-        remove_extraneous_path(path, file_type)?;
-        context.summary_mut().record_deletion();
-        context.record(LocalCopyRecord::new(
-            entry_relative,
-            LocalCopyAction::EntryDeleted,
-            0,
-            None,
-            Duration::default(),
-            None,
-        ));
-        context.register_progress();
-    }
-
-    if skipped_due_to_limit > 0 {
-        return Err(LocalCopyError::delete_limit_exceeded(skipped_due_to_limit));
-    }
-
-    Ok(())
-}
-
-fn remove_extraneous_path(path: PathBuf, file_type: fs::FileType) -> Result<(), LocalCopyError> {
-    let context = if file_type.is_dir() {
-        "remove extraneous directory"
-    } else {
-        "remove extraneous entry"
-    };
-
-    let result = if file_type.is_dir() {
-        fs::remove_dir_all(&path)
-    } else {
-        fs::remove_file(&path)
-    };
-
-    match result {
-        Ok(()) => Ok(()),
-        Err(error) if error.kind() == io::ErrorKind::NotFound => Ok(()),
-        Err(error) => Err(LocalCopyError::io(context, path, error)),
-    }
-}
-
-fn remove_source_entry_if_requested(
-    context: &mut CopyContext,
-    source: &Path,
-    record_path: Option<&Path>,
-    file_type: fs::FileType,
-) -> Result<(), LocalCopyError> {
-    if !context.remove_source_files_enabled() || context.mode().is_dry_run() {
-        return Ok(());
-    }
-
-    let source_type = match fs::symlink_metadata(source) {
-        Ok(metadata) => metadata.file_type(),
-        Err(_) => file_type,
-    };
-
-    if source_type.is_dir() {
-        return Ok(());
-    }
-
-    match fs::remove_file(source) {
-        Ok(()) => {
-            context.summary_mut().record_source_removed();
-            if let Some(path) = record_path {
-                context.record(LocalCopyRecord::new(
-                    path.to_path_buf(),
-                    LocalCopyAction::SourceRemoved,
-                    0,
-                    None,
-                    Duration::default(),
-                    None,
-                ));
-            }
-            context.register_progress();
-            Ok(())
-        }
-        Err(error) if error.kind() == io::ErrorKind::NotFound => Ok(()),
-        Err(error) => Err(LocalCopyError::io(
-            "remove source entry",
-            source.to_path_buf(),
-            error,
-        )),
-    }
-}
-
-fn symlink_target_is_safe(target: &Path, link_relative: &Path) -> bool {
-    if target.as_os_str().is_empty() || target.has_root() {
-        return false;
-    }
-
-    let mut seen_non_parent = false;
-    let mut last_was_parent = false;
-    let mut component_count = 0usize;
-
-    for component in target.components() {
-        match component {
-            Component::ParentDir => {
-                if seen_non_parent {
-                    return false;
-                }
-                last_was_parent = true;
-            }
-            Component::CurDir => {
-                seen_non_parent = true;
-                last_was_parent = false;
-            }
-            Component::Normal(_) => {
-                seen_non_parent = true;
-                last_was_parent = false;
-            }
-            Component::RootDir | Component::Prefix(_) => return false,
-        }
-        component_count = component_count.saturating_add(1);
-    }
-
-    if component_count > 1 && last_was_parent {
-        return false;
-    }
-
-    let mut depth: i64 = 0;
-    for component in link_relative.components() {
-        match component {
-            Component::ParentDir => depth = 0,
-            Component::CurDir => {}
-            Component::Normal(_) => depth += 1,
-            Component::RootDir | Component::Prefix(_) => depth = 0,
-        }
-    }
-
-    for component in target.components() {
-        match component {
-            Component::ParentDir => {
-                depth -= 1;
-                if depth < 0 {
-                    return false;
-                }
-            }
-            Component::CurDir => {}
-            Component::Normal(_) => depth += 1,
-            Component::RootDir | Component::Prefix(_) => return false,
-        }
-    }
-
-    true
-}
-
-fn copy_symlink(
-    context: &mut CopyContext,
-    source: &Path,
-    destination: &Path,
-    metadata: &fs::Metadata,
-    metadata_options: &MetadataOptions,
-    relative: Option<&Path>,
-) -> Result<(), LocalCopyError> {
-    context.enforce_timeout()?;
-    let mode = context.mode();
-    let file_type = metadata.file_type();
-    #[cfg(feature = "xattr")]
-    let preserve_xattrs = context.xattrs_enabled();
-    #[cfg(feature = "acl")]
-    let preserve_acls = context.acls_enabled();
-    let record_path = relative
-        .map(Path::to_path_buf)
-        .or_else(|| destination.file_name().map(PathBuf::from));
-    context.summary_mut().record_symlink_total();
-    let target = fs::read_link(source)
-        .map_err(|error| LocalCopyError::io("read symbolic link", source.to_path_buf(), error))?;
-
-    let safety_relative = relative
-        .map(Path::to_path_buf)
-        .or_else(|| {
-            destination
-                .strip_prefix(context.destination_root())
-                .ok()
-                .and_then(|path| (!path.as_os_str().is_empty()).then(|| path.to_path_buf()))
-        })
-        .or_else(|| destination.file_name().map(PathBuf::from))
-        .unwrap_or_else(|| destination.to_path_buf());
-
-    let unsafe_target =
-        context.safe_links_enabled() && !symlink_target_is_safe(&target, &safety_relative);
-
-    if unsafe_target {
-        if context.copy_unsafe_links_enabled() {
-            let target_metadata = follow_symlink_metadata(source)?;
-            let target_type = target_metadata.file_type();
-
-            if target_type.is_dir() {
-                let _kept = copy_directory_recursive(
-                    context,
-                    source,
-                    destination,
-                    &target_metadata,
-                    relative,
-                    None,
-                )?;
-                return Ok(());
-            }
-
-            if target_type.is_file() {
-                copy_file(context, source, destination, &target_metadata, relative)?;
-                return Ok(());
-            }
-
-            if is_fifo(&target_type) {
-                if !context.specials_enabled() {
-                    context.record_skipped_non_regular(record_path.as_deref());
-                    context.register_progress();
-                    return Ok(());
-                }
-                copy_fifo(
-                    context,
-                    source,
-                    destination,
-                    &target_metadata,
-                    metadata_options,
-                    relative,
-                )?;
-                return Ok(());
-            }
-
-            if is_device(&target_type) {
-                if !context.devices_enabled() {
-                    context.record_skipped_non_regular(record_path.as_deref());
-                    context.register_progress();
-                    return Ok(());
-                }
-                copy_device(
-                    context,
-                    source,
-                    destination,
-                    &target_metadata,
-                    metadata_options,
-                    relative,
-                )?;
-                return Ok(());
-            }
-
-            return Err(LocalCopyError::invalid_argument(
-                LocalCopyArgumentError::UnsupportedFileType,
-            ));
-        }
-
-        context.record_skipped_unsafe_symlink(record_path.as_deref(), metadata, target);
-        context.register_progress();
-        return Ok(());
-    }
-
-    if let Some(parent) = destination.parent() {
-        if !parent.as_os_str().is_empty() {
-            if mode.is_dry_run() {
-                match fs::symlink_metadata(parent) {
-                    Ok(existing) if !existing.file_type().is_dir() => {
-                        return Err(LocalCopyError::invalid_argument(
-                            LocalCopyArgumentError::ReplaceNonDirectoryWithDirectory,
-                        ));
-                    }
-                    Ok(_) => {}
-                    Err(error) if error.kind() == io::ErrorKind::NotFound => {}
-                    Err(error) => {
-                        return Err(LocalCopyError::io(
-                            "inspect existing destination",
-                            parent.to_path_buf(),
-                            error,
-                        ));
-                    }
-                }
-            } else {
-                fs::create_dir_all(parent).map_err(|error| {
-                    LocalCopyError::io("create parent directory", parent.to_path_buf(), error)
-                })?;
-                context.register_progress();
-            }
-        }
-    }
-
-    let mut destination_previously_existed = false;
-    match fs::symlink_metadata(destination) {
-        Ok(existing) => {
-            destination_previously_existed = true;
-            let file_type = existing.file_type();
-            if file_type.is_dir() {
-                return Err(LocalCopyError::invalid_argument(
-                    LocalCopyArgumentError::ReplaceDirectoryWithSymlink,
-                ));
-            }
-
-            if !mode.is_dry_run() {
-                context.backup_existing_entry(destination, relative, file_type)?;
-                remove_existing_destination(destination)?;
-            }
-        }
-        Err(error) if error.kind() == io::ErrorKind::NotFound => {}
-        Err(error) => {
-            return Err(LocalCopyError::io(
-                "inspect existing destination",
-                destination.to_path_buf(),
-                error,
-            ));
-        }
-    }
-
-    if mode.is_dry_run() {
-        context.summary_mut().record_symlink();
-        if let Some(path) = &record_path {
-            let metadata_snapshot =
-                LocalCopyMetadata::from_metadata(metadata, Some(target.clone()));
-            let total_bytes = Some(metadata_snapshot.len());
-            context.record(LocalCopyRecord::new(
-                path.clone(),
-                LocalCopyAction::SymlinkCopied,
-                0,
-                total_bytes,
-                Duration::default(),
-                Some(metadata_snapshot),
-            ));
-        }
-        context.register_progress();
-        remove_source_entry_if_requested(context, source, record_path.as_deref(), file_type)?;
-        return Ok(());
-    }
-
-    create_symlink(&target, source, destination).map_err(|error| {
-        LocalCopyError::io("create symbolic link", destination.to_path_buf(), error)
-    })?;
-
-    context.register_created_path(
-        destination,
-        CreatedEntryKind::Symlink,
-        destination_previously_existed,
-    );
-
-    let symlink_options = if context.omit_link_times_enabled() {
-        metadata_options.clone().preserve_times(false)
-    } else {
-        metadata_options.clone()
-    };
-    apply_symlink_metadata_with_options(destination, metadata, symlink_options)
-        .map_err(map_metadata_error)?;
-    #[cfg(feature = "xattr")]
-    sync_xattrs_if_requested(preserve_xattrs, mode, source, destination, false)?;
-    #[cfg(feature = "acl")]
-    sync_acls_if_requested(preserve_acls, mode, source, destination, false)?;
-
-    context.summary_mut().record_symlink();
-    if let Some(path) = &record_path {
-        let metadata_snapshot = LocalCopyMetadata::from_metadata(metadata, Some(target.clone()));
-        let total_bytes = Some(metadata_snapshot.len());
-        context.record(LocalCopyRecord::new(
-            path.clone(),
-            LocalCopyAction::SymlinkCopied,
-            0,
-            total_bytes,
-            Duration::default(),
-            Some(metadata_snapshot),
-        ));
-    }
-    context.register_progress();
-    remove_source_entry_if_requested(context, source, record_path.as_deref(), file_type)?;
-    Ok(())
-}
-
-fn ensure_destination_directory(
-    destination_path: &Path,
-    state: &mut DestinationState,
-    mode: LocalCopyExecution,
-) -> Result<(), LocalCopyError> {
-    if state.exists {
-        if !state.is_dir {
-            return Err(LocalCopyError::invalid_argument(
-                LocalCopyArgumentError::DestinationMustBeDirectory,
-            ));
-        }
-        return Ok(());
-    }
-
-    if mode.is_dry_run() {
-        state.exists = true;
-        state.is_dir = true;
-        return Ok(());
-    }
-
-    fs::create_dir_all(destination_path).map_err(|error| {
-        LocalCopyError::io(
-            "create destination directory",
-            destination_path.to_path_buf(),
-            error,
-        )
-    })?;
-    state.exists = true;
-    state.is_dir = true;
-    Ok(())
-}
-
-fn map_metadata_error(error: MetadataError) -> LocalCopyError {
-    let (context, path, source) = error.into_parts();
-    LocalCopyError::io(context, path, source)
-}
-
-fn read_directory_entries_sorted(path: &Path) -> Result<Vec<DirectoryEntry>, LocalCopyError> {
-    let mut entries = Vec::new();
-    let read_dir = fs::read_dir(path)
-        .map_err(|error| LocalCopyError::io("read directory", path.to_path_buf(), error))?;
-
-    for entry in read_dir {
-        let entry = entry.map_err(|error| {
-            LocalCopyError::io("read directory entry", path.to_path_buf(), error)
-        })?;
-        let entry_path = entry.path();
-        let metadata = fs::symlink_metadata(&entry_path).map_err(|error| {
-            LocalCopyError::io("inspect directory entry", entry_path.to_path_buf(), error)
-        })?;
-        entries.push(DirectoryEntry {
-            file_name: entry.file_name(),
-            path: entry_path,
-            metadata,
-        });
-    }
-
-    entries.sort_by(|a, b| compare_file_names(&a.file_name, &b.file_name));
-    Ok(entries)
-}
-
-fn compare_file_names(left: &OsStr, right: &OsStr) -> Ordering {
-    #[cfg(unix)]
-    {
-        use std::os::unix::ffi::OsStrExt;
-
-        left.as_bytes().cmp(right.as_bytes())
-    }
-
-    #[cfg(windows)]
-    {
-        use std::os::windows::ffi::OsStrExt;
-
-        let left_wide: Vec<u16> = left.encode_wide().collect();
-        let right_wide: Vec<u16> = right.encode_wide().collect();
-        left_wide.cmp(&right_wide)
-    }
-
-    #[cfg(not(any(unix, windows)))]
-    {
-        left.to_string_lossy().cmp(&right.to_string_lossy())
-    }
-}
-
-fn has_trailing_separator(path: &OsStr) -> bool {
-    #[cfg(unix)]
-    {
-        use std::os::unix::ffi::OsStrExt;
-
-        let bytes = path.as_bytes();
-        !bytes.is_empty() && bytes.ends_with(b"/")
-    }
-
-    #[cfg(windows)]
-    {
-        use std::os::windows::ffi::OsStrExt;
-
-        let mut last_nonzero = None;
-        for ch in path.encode_wide() {
-            if ch != 0 {
-                last_nonzero = Some(ch);
-            }
-        }
-
-        last_nonzero.is_some_and(|ch| ch == b'/' as u16 || ch == b'\\' as u16)
-    }
-
-    #[cfg(not(any(unix, windows)))]
-    {
-        let text = path.to_string_lossy();
-        text.ends_with('/') || text.ends_with('\\')
-    }
-}
-
-fn is_fifo(file_type: &fs::FileType) -> bool {
-    #[cfg(unix)]
-    {
-        use std::os::unix::fs::FileTypeExt;
-
-        file_type.is_fifo()
-    }
-
-    #[cfg(not(unix))]
-    {
-        let _ = file_type;
-        false
-    }
-}
-
-fn is_device(file_type: &fs::FileType) -> bool {
-    #[cfg(unix)]
-    {
-        use std::os::unix::fs::FileTypeExt;
-
-        file_type.is_char_device() || file_type.is_block_device()
-    }
-
-    #[cfg(not(unix))]
-    {
-        let _ = file_type;
-        false
-    }
-}
-
-#[cfg(windows)]
-fn operand_has_windows_prefix(path: &OsStr) -> bool {
-    use std::os::windows::ffi::OsStrExt;
-
-    const COLON: u16 = b':' as u16;
-    const QUESTION: u16 = b'?' as u16;
-    const DOT: u16 = b'.' as u16;
-    const SLASH: u16 = b'/' as u16;
-    const BACKSLASH: u16 = b'\\' as u16;
-
-    fn is_ascii_alpha(unit: u16) -> bool {
-        (unit >= b'a' as u16 && unit <= b'z' as u16) || (unit >= b'A' as u16 && unit <= b'Z' as u16)
-    }
-
-    fn is_separator(unit: u16) -> bool {
-        unit == SLASH || unit == BACKSLASH
-    }
-
-    let units: Vec<u16> = path.encode_wide().collect();
-    if units.is_empty() {
-        return false;
-    }
-
-    if units.len() >= 4
-        && is_separator(units[0])
-        && is_separator(units[1])
-        && (units[2] == QUESTION || units[2] == DOT)
-        && is_separator(units[3])
-    {
-        return true;
-    }
-
-    if units.len() >= 2 && is_separator(units[0]) && is_separator(units[1]) {
-        return true;
-    }
-
-    if units.len() >= 2 && is_ascii_alpha(units[0]) && units[1] == COLON {
-        return true;
-    }
-
-    false
-}
-
-fn operand_is_remote(path: &OsStr) -> bool {
-    let text = path.to_string_lossy();
-
-    if text.starts_with("rsync://") {
-        return true;
-    }
-
-    if text.contains("::") {
-        return true;
-    }
-
-    if let Some(colon_index) = text.find(':') {
-        #[cfg(windows)]
-        if operand_has_windows_prefix(path) {
-            return false;
-        }
-
-        let after = &text[colon_index + 1..];
-        if after.starts_with(':') {
-            return true;
-        }
-
-        let before = &text[..colon_index];
-        if before.contains('/') || before.contains('\\') {
-            return false;
-        }
-
-        if colon_index == 1 && before.chars().all(|ch| ch.is_ascii_alphabetic()) {
-            return false;
-        }
-
-        return true;
-    }
-
-    false
-}
-
-#[cfg(all(test, windows))]
-mod windows_operand_detection {
-    use super::operand_is_remote;
-    use std::ffi::OsStr;
-
-    #[test]
-    fn drive_letter_paths_are_local() {
-        assert!(!operand_is_remote(OsStr::new(r"C:\\tmp\\file.txt")));
-        assert!(!operand_is_remote(OsStr::new(r"d:relative\\path")));
-    }
-
-    #[test]
-    fn extended_prefixes_are_local() {
-        assert!(!operand_is_remote(OsStr::new(r"\\\\?\\C:\\tmp\\file.txt")));
-        assert!(!operand_is_remote(OsStr::new(
-            r"\\\\?\\UNC\\server\\share\\file.txt"
-        )));
-        assert!(!operand_is_remote(OsStr::new(r"\\\\.\\pipe\\rsync")));
-    }
-
-    #[test]
-    fn unc_and_forward_slash_paths_are_local() {
-        assert!(!operand_is_remote(OsStr::new(
-            r"\\\\server\\share\\file.txt"
-        )));
-        assert!(!operand_is_remote(OsStr::new("//server/share/file.txt")));
-    }
-
-    #[test]
-    fn remote_operands_remain_remote() {
-        assert!(operand_is_remote(OsStr::new("host:path")));
-        assert!(operand_is_remote(OsStr::new("user@host:path")));
-        assert!(operand_is_remote(OsStr::new("host::module")));
-        assert!(operand_is_remote(OsStr::new("rsync://example.com/module")));
-    }
-}
-
-#[cfg(unix)]
-fn create_symlink(target: &Path, _source: &Path, destination: &Path) -> io::Result<()> {
-    use std::os::unix::fs::symlink;
-
-    symlink(target, destination)
-}
-
-#[cfg(windows)]
-fn create_symlink(target: &Path, source: &Path, destination: &Path) -> io::Result<()> {
-    use std::os::windows::fs::{symlink_dir, symlink_file};
-
-    match source.metadata() {
-        Ok(metadata) if metadata.file_type().is_dir() => symlink_dir(target, destination),
-        Ok(_) => symlink_file(target, destination),
-        Err(_) => symlink_file(target, destination),
-    }
+#[cfg(test)]
+pub(crate) fn with_hard_link_override<F, R>(override_fn: F, action: impl FnOnce() -> R) -> R
+where
+    F: Fn(&Path, &Path) -> std::io::Result<()> + 'static,
+{
+    overrides::with_hard_link_override(override_fn, action)
+}
+
+#[cfg(test)]
+pub(crate) fn with_device_id_override<F, R>(override_fn: F, action: impl FnOnce() -> R) -> R
+where
+    F: Fn(&Path, &fs::Metadata) -> Option<u64> + 'static,
+{
+    overrides::with_device_id_override(override_fn, action)
 }
 
 #[cfg(test)]
